Module aws_sdk_iot::client::fluent_builders

Utilities to ergonomically construct a request to the service.

Fluent builders are created through the Client by calling one if its operation methods. After parameters are set using the builder methods, the send method can be called to initiate the request.

Structs

AcceptCertificateTransfer

Fluent builder constructing a request to AcceptCertificateTransfer.

AddThingToBillingGroup

Fluent builder constructing a request to AddThingToBillingGroup.

AddThingToThingGroup

Fluent builder constructing a request to AddThingToThingGroup.

AssociateTargetsWithJob

Fluent builder constructing a request to AssociateTargetsWithJob.

AttachPolicy

Fluent builder constructing a request to AttachPolicy.

AttachPrincipalPolicy

Fluent builder constructing a request to AttachPrincipalPolicy.

AttachSecurityProfile

Fluent builder constructing a request to AttachSecurityProfile.

AttachThingPrincipal

Fluent builder constructing a request to AttachThingPrincipal.

CancelAuditMitigationActionsTask

Fluent builder constructing a request to CancelAuditMitigationActionsTask.

CancelAuditTask

Fluent builder constructing a request to CancelAuditTask.

CancelCertificateTransfer

Fluent builder constructing a request to CancelCertificateTransfer.

CancelDetectMitigationActionsTask

Fluent builder constructing a request to CancelDetectMitigationActionsTask.

CancelJob

Fluent builder constructing a request to CancelJob.

CancelJobExecution

Fluent builder constructing a request to CancelJobExecution.

ClearDefaultAuthorizer

Fluent builder constructing a request to ClearDefaultAuthorizer.

ConfirmTopicRuleDestination

Fluent builder constructing a request to ConfirmTopicRuleDestination.

CreateAuditSuppression

Fluent builder constructing a request to CreateAuditSuppression.

CreateAuthorizer

Fluent builder constructing a request to CreateAuthorizer.

CreateBillingGroup

Fluent builder constructing a request to CreateBillingGroup.

CreateCertificateFromCsr

Fluent builder constructing a request to CreateCertificateFromCsr.

CreateCustomMetric

Fluent builder constructing a request to CreateCustomMetric.

CreateDimension

Fluent builder constructing a request to CreateDimension.

CreateDomainConfiguration

Fluent builder constructing a request to CreateDomainConfiguration.

CreateDynamicThingGroup

Fluent builder constructing a request to CreateDynamicThingGroup.

CreateFleetMetric

Fluent builder constructing a request to CreateFleetMetric.

CreateJob

Fluent builder constructing a request to CreateJob.

CreateJobTemplate

Fluent builder constructing a request to CreateJobTemplate.

CreateKeysAndCertificate

Fluent builder constructing a request to CreateKeysAndCertificate.

CreateMitigationAction

Fluent builder constructing a request to CreateMitigationAction.

CreateOTAUpdate

Fluent builder constructing a request to CreateOTAUpdate.

CreatePolicy

Fluent builder constructing a request to CreatePolicy.

CreatePolicyVersion

Fluent builder constructing a request to CreatePolicyVersion.

CreateProvisioningClaim

Fluent builder constructing a request to CreateProvisioningClaim.

CreateProvisioningTemplate

Fluent builder constructing a request to CreateProvisioningTemplate.

CreateProvisioningTemplateVersion

Fluent builder constructing a request to CreateProvisioningTemplateVersion.

CreateRoleAlias

Fluent builder constructing a request to CreateRoleAlias.

CreateScheduledAudit

Fluent builder constructing a request to CreateScheduledAudit.

CreateSecurityProfile

Fluent builder constructing a request to CreateSecurityProfile.

CreateStream

Fluent builder constructing a request to CreateStream.

CreateThing

Fluent builder constructing a request to CreateThing.

CreateThingGroup

Fluent builder constructing a request to CreateThingGroup.

CreateThingType

Fluent builder constructing a request to CreateThingType.

CreateTopicRule

Fluent builder constructing a request to CreateTopicRule.

CreateTopicRuleDestination

Fluent builder constructing a request to CreateTopicRuleDestination.

DeleteAccountAuditConfiguration

Fluent builder constructing a request to DeleteAccountAuditConfiguration.

DeleteAuditSuppression

Fluent builder constructing a request to DeleteAuditSuppression.

DeleteAuthorizer

Fluent builder constructing a request to DeleteAuthorizer.

DeleteBillingGroup

Fluent builder constructing a request to DeleteBillingGroup.

DeleteCACertificate

Fluent builder constructing a request to DeleteCACertificate.

DeleteCertificate

Fluent builder constructing a request to DeleteCertificate.

DeleteCustomMetric

Fluent builder constructing a request to DeleteCustomMetric.

DeleteDimension

Fluent builder constructing a request to DeleteDimension.

DeleteDomainConfiguration

Fluent builder constructing a request to DeleteDomainConfiguration.

DeleteDynamicThingGroup

Fluent builder constructing a request to DeleteDynamicThingGroup.

DeleteFleetMetric

Fluent builder constructing a request to DeleteFleetMetric.

DeleteJob

Fluent builder constructing a request to DeleteJob.

DeleteJobExecution

Fluent builder constructing a request to DeleteJobExecution.

DeleteJobTemplate

Fluent builder constructing a request to DeleteJobTemplate.

DeleteMitigationAction

Fluent builder constructing a request to DeleteMitigationAction.

DeleteOTAUpdate

Fluent builder constructing a request to DeleteOTAUpdate.

DeletePolicy

Fluent builder constructing a request to DeletePolicy.

DeletePolicyVersion

Fluent builder constructing a request to DeletePolicyVersion.

DeleteProvisioningTemplate

Fluent builder constructing a request to DeleteProvisioningTemplate.

DeleteProvisioningTemplateVersion

Fluent builder constructing a request to DeleteProvisioningTemplateVersion.

DeleteRegistrationCode

Fluent builder constructing a request to DeleteRegistrationCode.

DeleteRoleAlias

Fluent builder constructing a request to DeleteRoleAlias.

DeleteScheduledAudit

Fluent builder constructing a request to DeleteScheduledAudit.

DeleteSecurityProfile

Fluent builder constructing a request to DeleteSecurityProfile.

DeleteStream

Fluent builder constructing a request to DeleteStream.

DeleteThing

Fluent builder constructing a request to DeleteThing.

DeleteThingGroup

Fluent builder constructing a request to DeleteThingGroup.

DeleteThingType

Fluent builder constructing a request to DeleteThingType.

DeleteTopicRule

Fluent builder constructing a request to DeleteTopicRule.

DeleteTopicRuleDestination

Fluent builder constructing a request to DeleteTopicRuleDestination.

DeleteV2LoggingLevel

Fluent builder constructing a request to DeleteV2LoggingLevel.

DeprecateThingType

Fluent builder constructing a request to DeprecateThingType.

DescribeAccountAuditConfiguration

Fluent builder constructing a request to DescribeAccountAuditConfiguration.

DescribeAuditFinding

Fluent builder constructing a request to DescribeAuditFinding.

DescribeAuditMitigationActionsTask

Fluent builder constructing a request to DescribeAuditMitigationActionsTask.

DescribeAuditSuppression

Fluent builder constructing a request to DescribeAuditSuppression.

DescribeAuditTask

Fluent builder constructing a request to DescribeAuditTask.

DescribeAuthorizer

Fluent builder constructing a request to DescribeAuthorizer.

DescribeBillingGroup

Fluent builder constructing a request to DescribeBillingGroup.

DescribeCACertificate

Fluent builder constructing a request to DescribeCACertificate.

DescribeCertificate

Fluent builder constructing a request to DescribeCertificate.

DescribeCustomMetric

Fluent builder constructing a request to DescribeCustomMetric.

DescribeDefaultAuthorizer

Fluent builder constructing a request to DescribeDefaultAuthorizer.

DescribeDetectMitigationActionsTask

Fluent builder constructing a request to DescribeDetectMitigationActionsTask.

DescribeDimension

Fluent builder constructing a request to DescribeDimension.

DescribeDomainConfiguration

Fluent builder constructing a request to DescribeDomainConfiguration.

DescribeEndpoint

Fluent builder constructing a request to DescribeEndpoint.

DescribeEventConfigurations

Fluent builder constructing a request to DescribeEventConfigurations.

DescribeFleetMetric

Fluent builder constructing a request to DescribeFleetMetric.

DescribeIndex

Fluent builder constructing a request to DescribeIndex.

DescribeJob

Fluent builder constructing a request to DescribeJob.

DescribeJobExecution

Fluent builder constructing a request to DescribeJobExecution.

DescribeJobTemplate

Fluent builder constructing a request to DescribeJobTemplate.

DescribeMitigationAction

Fluent builder constructing a request to DescribeMitigationAction.

DescribeProvisioningTemplate

Fluent builder constructing a request to DescribeProvisioningTemplate.

DescribeProvisioningTemplateVersion

Fluent builder constructing a request to DescribeProvisioningTemplateVersion.

DescribeRoleAlias

Fluent builder constructing a request to DescribeRoleAlias.

DescribeScheduledAudit

Fluent builder constructing a request to DescribeScheduledAudit.

DescribeSecurityProfile

Fluent builder constructing a request to DescribeSecurityProfile.

DescribeStream

Fluent builder constructing a request to DescribeStream.

DescribeThing

Fluent builder constructing a request to DescribeThing.

DescribeThingGroup

Fluent builder constructing a request to DescribeThingGroup.

DescribeThingRegistrationTask

Fluent builder constructing a request to DescribeThingRegistrationTask.

DescribeThingType

Fluent builder constructing a request to DescribeThingType.

DetachPolicy

Fluent builder constructing a request to DetachPolicy.

DetachPrincipalPolicy

Fluent builder constructing a request to DetachPrincipalPolicy.

DetachSecurityProfile

Fluent builder constructing a request to DetachSecurityProfile.

DetachThingPrincipal

Fluent builder constructing a request to DetachThingPrincipal.

DisableTopicRule

Fluent builder constructing a request to DisableTopicRule.

EnableTopicRule

Fluent builder constructing a request to EnableTopicRule.

GetBehaviorModelTrainingSummaries

Fluent builder constructing a request to GetBehaviorModelTrainingSummaries.

GetBucketsAggregation

Fluent builder constructing a request to GetBucketsAggregation.

GetCardinality

Fluent builder constructing a request to GetCardinality.

GetEffectivePolicies

Fluent builder constructing a request to GetEffectivePolicies.

GetIndexingConfiguration

Fluent builder constructing a request to GetIndexingConfiguration.

GetJobDocument

Fluent builder constructing a request to GetJobDocument.

GetLoggingOptions

Fluent builder constructing a request to GetLoggingOptions.

GetOTAUpdate

Fluent builder constructing a request to GetOTAUpdate.

GetPercentiles

Fluent builder constructing a request to GetPercentiles.

GetPolicy

Fluent builder constructing a request to GetPolicy.

GetPolicyVersion

Fluent builder constructing a request to GetPolicyVersion.

GetRegistrationCode

Fluent builder constructing a request to GetRegistrationCode.

GetStatistics

Fluent builder constructing a request to GetStatistics.

GetTopicRule

Fluent builder constructing a request to GetTopicRule.

GetTopicRuleDestination

Fluent builder constructing a request to GetTopicRuleDestination.

GetV2LoggingOptions

Fluent builder constructing a request to GetV2LoggingOptions.

ListActiveViolations

Fluent builder constructing a request to ListActiveViolations.

ListAttachedPolicies

Fluent builder constructing a request to ListAttachedPolicies.

ListAuditFindings

Fluent builder constructing a request to ListAuditFindings.

ListAuditMitigationActionsExecutions

Fluent builder constructing a request to ListAuditMitigationActionsExecutions.

ListAuditMitigationActionsTasks

Fluent builder constructing a request to ListAuditMitigationActionsTasks.

ListAuditSuppressions

Fluent builder constructing a request to ListAuditSuppressions.

ListAuditTasks

Fluent builder constructing a request to ListAuditTasks.

ListAuthorizers

Fluent builder constructing a request to ListAuthorizers.

ListBillingGroups

Fluent builder constructing a request to ListBillingGroups.

ListCACertificates

Fluent builder constructing a request to ListCACertificates.

ListCertificates

Fluent builder constructing a request to ListCertificates.

ListCertificatesByCA

Fluent builder constructing a request to ListCertificatesByCA.

ListCustomMetrics

Fluent builder constructing a request to ListCustomMetrics.

ListDetectMitigationActionsExecutions

Fluent builder constructing a request to ListDetectMitigationActionsExecutions.

ListDetectMitigationActionsTasks

Fluent builder constructing a request to ListDetectMitigationActionsTasks.

ListDimensions

Fluent builder constructing a request to ListDimensions.

ListDomainConfigurations

Fluent builder constructing a request to ListDomainConfigurations.

ListFleetMetrics

Fluent builder constructing a request to ListFleetMetrics.

ListIndices

Fluent builder constructing a request to ListIndices.

ListJobExecutionsForJob

Fluent builder constructing a request to ListJobExecutionsForJob.

ListJobExecutionsForThing

Fluent builder constructing a request to ListJobExecutionsForThing.

ListJobTemplates

Fluent builder constructing a request to ListJobTemplates.

ListJobs

Fluent builder constructing a request to ListJobs.

ListMitigationActions

Fluent builder constructing a request to ListMitigationActions.

ListOTAUpdates

Fluent builder constructing a request to ListOTAUpdates.

ListOutgoingCertificates

Fluent builder constructing a request to ListOutgoingCertificates.

ListPolicies

Fluent builder constructing a request to ListPolicies.

ListPolicyPrincipals

Fluent builder constructing a request to ListPolicyPrincipals.

ListPolicyVersions

Fluent builder constructing a request to ListPolicyVersions.

ListPrincipalPolicies

Fluent builder constructing a request to ListPrincipalPolicies.

ListPrincipalThings

Fluent builder constructing a request to ListPrincipalThings.

ListProvisioningTemplateVersions

Fluent builder constructing a request to ListProvisioningTemplateVersions.

ListProvisioningTemplates

Fluent builder constructing a request to ListProvisioningTemplates.

ListRoleAliases

Fluent builder constructing a request to ListRoleAliases.

ListScheduledAudits

Fluent builder constructing a request to ListScheduledAudits.

ListSecurityProfiles

Fluent builder constructing a request to ListSecurityProfiles.

ListSecurityProfilesForTarget

Fluent builder constructing a request to ListSecurityProfilesForTarget.

ListStreams

Fluent builder constructing a request to ListStreams.

ListTagsForResource

Fluent builder constructing a request to ListTagsForResource.

ListTargetsForPolicy

Fluent builder constructing a request to ListTargetsForPolicy.

ListTargetsForSecurityProfile

Fluent builder constructing a request to ListTargetsForSecurityProfile.

ListThingGroups

Fluent builder constructing a request to ListThingGroups.

ListThingGroupsForThing

Fluent builder constructing a request to ListThingGroupsForThing.

ListThingPrincipals

Fluent builder constructing a request to ListThingPrincipals.

ListThingRegistrationTaskReports

Fluent builder constructing a request to ListThingRegistrationTaskReports.

ListThingRegistrationTasks

Fluent builder constructing a request to ListThingRegistrationTasks.

ListThingTypes

Fluent builder constructing a request to ListThingTypes.

ListThings

Fluent builder constructing a request to ListThings.

ListThingsInBillingGroup

Fluent builder constructing a request to ListThingsInBillingGroup.

ListThingsInThingGroup

Fluent builder constructing a request to ListThingsInThingGroup.

ListTopicRuleDestinations

Fluent builder constructing a request to ListTopicRuleDestinations.

ListTopicRules

Fluent builder constructing a request to ListTopicRules.

ListV2LoggingLevels

Fluent builder constructing a request to ListV2LoggingLevels.

ListViolationEvents

Fluent builder constructing a request to ListViolationEvents.

PutVerificationStateOnViolation

Fluent builder constructing a request to PutVerificationStateOnViolation.

RegisterCACertificate

Fluent builder constructing a request to RegisterCACertificate.

RegisterCertificate

Fluent builder constructing a request to RegisterCertificate.

RegisterCertificateWithoutCA

Fluent builder constructing a request to RegisterCertificateWithoutCA.

RegisterThing

Fluent builder constructing a request to RegisterThing.

RejectCertificateTransfer

Fluent builder constructing a request to RejectCertificateTransfer.

RemoveThingFromBillingGroup

Fluent builder constructing a request to RemoveThingFromBillingGroup.

RemoveThingFromThingGroup

Fluent builder constructing a request to RemoveThingFromThingGroup.

ReplaceTopicRule

Fluent builder constructing a request to ReplaceTopicRule.

SearchIndex

Fluent builder constructing a request to SearchIndex.

SetDefaultAuthorizer

Fluent builder constructing a request to SetDefaultAuthorizer.

SetDefaultPolicyVersion

Fluent builder constructing a request to SetDefaultPolicyVersion.

SetLoggingOptions

Fluent builder constructing a request to SetLoggingOptions.

SetV2LoggingLevel

Fluent builder constructing a request to SetV2LoggingLevel.

SetV2LoggingOptions

Fluent builder constructing a request to SetV2LoggingOptions.

StartAuditMitigationActionsTask

Fluent builder constructing a request to StartAuditMitigationActionsTask.

StartDetectMitigationActionsTask

Fluent builder constructing a request to StartDetectMitigationActionsTask.

StartOnDemandAuditTask

Fluent builder constructing a request to StartOnDemandAuditTask.

StartThingRegistrationTask

Fluent builder constructing a request to StartThingRegistrationTask.

StopThingRegistrationTask

Fluent builder constructing a request to StopThingRegistrationTask.

TagResource

Fluent builder constructing a request to TagResource.

TestAuthorization

Fluent builder constructing a request to TestAuthorization.

TestInvokeAuthorizer

Fluent builder constructing a request to TestInvokeAuthorizer.

TransferCertificate

Fluent builder constructing a request to TransferCertificate.

UntagResource

Fluent builder constructing a request to UntagResource.

UpdateAccountAuditConfiguration

Fluent builder constructing a request to UpdateAccountAuditConfiguration.

UpdateAuditSuppression

Fluent builder constructing a request to UpdateAuditSuppression.

UpdateAuthorizer

Fluent builder constructing a request to UpdateAuthorizer.

UpdateBillingGroup

Fluent builder constructing a request to UpdateBillingGroup.

UpdateCACertificate

Fluent builder constructing a request to UpdateCACertificate.

UpdateCertificate

Fluent builder constructing a request to UpdateCertificate.

UpdateCustomMetric

Fluent builder constructing a request to UpdateCustomMetric.

UpdateDimension

Fluent builder constructing a request to UpdateDimension.

UpdateDomainConfiguration

Fluent builder constructing a request to UpdateDomainConfiguration.

UpdateDynamicThingGroup

Fluent builder constructing a request to UpdateDynamicThingGroup.

UpdateEventConfigurations

Fluent builder constructing a request to UpdateEventConfigurations.

UpdateFleetMetric

Fluent builder constructing a request to UpdateFleetMetric.

UpdateIndexingConfiguration

Fluent builder constructing a request to UpdateIndexingConfiguration.

UpdateJob

Fluent builder constructing a request to UpdateJob.

UpdateMitigationAction

Fluent builder constructing a request to UpdateMitigationAction.

UpdateProvisioningTemplate

Fluent builder constructing a request to UpdateProvisioningTemplate.

UpdateRoleAlias

Fluent builder constructing a request to UpdateRoleAlias.

UpdateScheduledAudit

Fluent builder constructing a request to UpdateScheduledAudit.

UpdateSecurityProfile

Fluent builder constructing a request to UpdateSecurityProfile.

UpdateStream

Fluent builder constructing a request to UpdateStream.

UpdateThing

Fluent builder constructing a request to UpdateThing.

UpdateThingGroup

Fluent builder constructing a request to UpdateThingGroup.

UpdateThingGroupsForThing

Fluent builder constructing a request to UpdateThingGroupsForThing.

UpdateTopicRuleDestination

Fluent builder constructing a request to UpdateTopicRuleDestination.

ValidateSecurityProfileBehaviors

Fluent builder constructing a request to ValidateSecurityProfileBehaviors.