Struct aws_sdk_iot::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSIotService.

This client allows ergonomic access to a AWSIotService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

pub fn conf(&self) -> &Config

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_certificate_transfer(&self) -> AcceptCertificateTransfer<C, M, R>

pub fn add_thing_to_billing_group(&self) -> AddThingToBillingGroup<C, M, R>

pub fn add_thing_to_thing_group(&self) -> AddThingToThingGroup<C, M, R>

pub fn associate_targets_with_job(&self) -> AssociateTargetsWithJob<C, M, R>

pub fn attach_policy(&self) -> AttachPolicy<C, M, R>

pub fn attach_principal_policy(&self) -> AttachPrincipalPolicy<C, M, R>

pub fn attach_security_profile(&self) -> AttachSecurityProfile<C, M, R>

pub fn attach_thing_principal(&self) -> AttachThingPrincipal<C, M, R>

pub fn cancel_audit_mitigation_actions_task(
&self
) -> CancelAuditMitigationActionsTask<C, M, R>

pub fn cancel_audit_task(&self) -> CancelAuditTask<C, M, R>

pub fn cancel_certificate_transfer(&self) -> CancelCertificateTransfer<C, M, R>

pub fn cancel_detect_mitigation_actions_task(
&self
) -> CancelDetectMitigationActionsTask<C, M, R>

pub fn cancel_job(&self) -> CancelJob<C, M, R>

pub fn cancel_job_execution(&self) -> CancelJobExecution<C, M, R>

pub fn clear_default_authorizer(&self) -> ClearDefaultAuthorizer<C, M, R>

pub fn confirm_topic_rule_destination(
&self
) -> ConfirmTopicRuleDestination<C, M, R>

pub fn create_audit_suppression(&self) -> CreateAuditSuppression<C, M, R>

pub fn create_authorizer(&self) -> CreateAuthorizer<C, M, R>

pub fn create_billing_group(&self) -> CreateBillingGroup<C, M, R>

pub fn create_certificate_from_csr(&self) -> CreateCertificateFromCsr<C, M, R>

pub fn create_custom_metric(&self) -> CreateCustomMetric<C, M, R>

pub fn create_dimension(&self) -> CreateDimension<C, M, R>

pub fn create_domain_configuration(&self) -> CreateDomainConfiguration<C, M, R>

pub fn create_dynamic_thing_group(&self) -> CreateDynamicThingGroup<C, M, R>

pub fn create_fleet_metric(&self) -> CreateFleetMetric<C, M, R>

pub fn create_job(&self) -> CreateJob<C, M, R>

pub fn create_job_template(&self) -> CreateJobTemplate<C, M, R>

pub fn create_keys_and_certificate(&self) -> CreateKeysAndCertificate<C, M, R>

pub fn create_mitigation_action(&self) -> CreateMitigationAction<C, M, R>

pub fn create_ota_update(&self) -> CreateOTAUpdate<C, M, R>

pub fn create_policy(&self) -> CreatePolicy<C, M, R>

pub fn create_policy_version(&self) -> CreatePolicyVersion<C, M, R>

pub fn create_provisioning_claim(&self) -> CreateProvisioningClaim<C, M, R>

pub fn create_provisioning_template(
&self
) -> CreateProvisioningTemplate<C, M, R>

pub fn create_provisioning_template_version(
&self
) -> CreateProvisioningTemplateVersion<C, M, R>

pub fn create_role_alias(&self) -> CreateRoleAlias<C, M, R>

pub fn create_scheduled_audit(&self) -> CreateScheduledAudit<C, M, R>

pub fn create_security_profile(&self) -> CreateSecurityProfile<C, M, R>

pub fn create_stream(&self) -> CreateStream<C, M, R>

pub fn create_thing(&self) -> CreateThing<C, M, R>

pub fn create_thing_group(&self) -> CreateThingGroup<C, M, R>

pub fn create_thing_type(&self) -> CreateThingType<C, M, R>

pub fn create_topic_rule(&self) -> CreateTopicRule<C, M, R>

pub fn create_topic_rule_destination(
&self
) -> CreateTopicRuleDestination<C, M, R>

pub fn delete_account_audit_configuration(
&self
) -> DeleteAccountAuditConfiguration<C, M, R>

pub fn delete_audit_suppression(&self) -> DeleteAuditSuppression<C, M, R>

pub fn delete_authorizer(&self) -> DeleteAuthorizer<C, M, R>

pub fn delete_billing_group(&self) -> DeleteBillingGroup<C, M, R>

pub fn delete_ca_certificate(&self) -> DeleteCACertificate<C, M, R>

pub fn delete_certificate(&self) -> DeleteCertificate<C, M, R>

pub fn delete_custom_metric(&self) -> DeleteCustomMetric<C, M, R>

pub fn delete_dimension(&self) -> DeleteDimension<C, M, R>

pub fn delete_domain_configuration(&self) -> DeleteDomainConfiguration<C, M, R>

pub fn delete_dynamic_thing_group(&self) -> DeleteDynamicThingGroup<C, M, R>

pub fn delete_fleet_metric(&self) -> DeleteFleetMetric<C, M, R>

pub fn delete_job(&self) -> DeleteJob<C, M, R>

pub fn delete_job_execution(&self) -> DeleteJobExecution<C, M, R>

pub fn delete_job_template(&self) -> DeleteJobTemplate<C, M, R>

pub fn delete_mitigation_action(&self) -> DeleteMitigationAction<C, M, R>

pub fn delete_ota_update(&self) -> DeleteOTAUpdate<C, M, R>

pub fn delete_policy(&self) -> DeletePolicy<C, M, R>

pub fn delete_policy_version(&self) -> DeletePolicyVersion<C, M, R>

pub fn delete_provisioning_template(
&self
) -> DeleteProvisioningTemplate<C, M, R>

pub fn delete_provisioning_template_version(
&self
) -> DeleteProvisioningTemplateVersion<C, M, R>

pub fn delete_registration_code(&self) -> DeleteRegistrationCode<C, M, R>

pub fn delete_role_alias(&self) -> DeleteRoleAlias<C, M, R>

pub fn delete_scheduled_audit(&self) -> DeleteScheduledAudit<C, M, R>

pub fn delete_security_profile(&self) -> DeleteSecurityProfile<C, M, R>

pub fn delete_stream(&self) -> DeleteStream<C, M, R>

pub fn delete_thing(&self) -> DeleteThing<C, M, R>

pub fn delete_thing_group(&self) -> DeleteThingGroup<C, M, R>

pub fn delete_thing_type(&self) -> DeleteThingType<C, M, R>

pub fn delete_topic_rule(&self) -> DeleteTopicRule<C, M, R>

pub fn delete_topic_rule_destination(
&self
) -> DeleteTopicRuleDestination<C, M, R>

pub fn delete_v2_logging_level(&self) -> DeleteV2LoggingLevel<C, M, R>

pub fn deprecate_thing_type(&self) -> DeprecateThingType<C, M, R>

pub fn describe_account_audit_configuration(
&self
) -> DescribeAccountAuditConfiguration<C, M, R>

pub fn describe_audit_finding(&self) -> DescribeAuditFinding<C, M, R>

pub fn describe_audit_mitigation_actions_task(
&self
) -> DescribeAuditMitigationActionsTask<C, M, R>

pub fn describe_audit_suppression(&self) -> DescribeAuditSuppression<C, M, R>

pub fn describe_audit_task(&self) -> DescribeAuditTask<C, M, R>

pub fn describe_authorizer(&self) -> DescribeAuthorizer<C, M, R>

pub fn describe_billing_group(&self) -> DescribeBillingGroup<C, M, R>

pub fn describe_ca_certificate(&self) -> DescribeCACertificate<C, M, R>

pub fn describe_certificate(&self) -> DescribeCertificate<C, M, R>

pub fn describe_custom_metric(&self) -> DescribeCustomMetric<C, M, R>

pub fn describe_default_authorizer(&self) -> DescribeDefaultAuthorizer<C, M, R>

pub fn describe_detect_mitigation_actions_task(
&self
) -> DescribeDetectMitigationActionsTask<C, M, R>

pub fn describe_dimension(&self) -> DescribeDimension<C, M, R>

pub fn describe_domain_configuration(
&self
) -> DescribeDomainConfiguration<C, M, R>

pub fn describe_endpoint(&self) -> DescribeEndpoint<C, M, R>

pub fn describe_event_configurations(
&self
) -> DescribeEventConfigurations<C, M, R>

pub fn describe_fleet_metric(&self) -> DescribeFleetMetric<C, M, R>

pub fn describe_index(&self) -> DescribeIndex<C, M, R>

pub fn describe_job(&self) -> DescribeJob<C, M, R>

pub fn describe_job_execution(&self) -> DescribeJobExecution<C, M, R>

pub fn describe_job_template(&self) -> DescribeJobTemplate<C, M, R>

pub fn describe_mitigation_action(&self) -> DescribeMitigationAction<C, M, R>

pub fn describe_provisioning_template(
&self
) -> DescribeProvisioningTemplate<C, M, R>

pub fn describe_provisioning_template_version(
&self
) -> DescribeProvisioningTemplateVersion<C, M, R>

pub fn describe_role_alias(&self) -> DescribeRoleAlias<C, M, R>

pub fn describe_scheduled_audit(&self) -> DescribeScheduledAudit<C, M, R>

pub fn describe_security_profile(&self) -> DescribeSecurityProfile<C, M, R>

pub fn describe_stream(&self) -> DescribeStream<C, M, R>

pub fn describe_thing(&self) -> DescribeThing<C, M, R>

pub fn describe_thing_group(&self) -> DescribeThingGroup<C, M, R>

pub fn describe_thing_registration_task(
&self
) -> DescribeThingRegistrationTask<C, M, R>

pub fn describe_thing_type(&self) -> DescribeThingType<C, M, R>

pub fn detach_policy(&self) -> DetachPolicy<C, M, R>

pub fn detach_principal_policy(&self) -> DetachPrincipalPolicy<C, M, R>

pub fn detach_security_profile(&self) -> DetachSecurityProfile<C, M, R>

pub fn detach_thing_principal(&self) -> DetachThingPrincipal<C, M, R>

pub fn disable_topic_rule(&self) -> DisableTopicRule<C, M, R>

pub fn enable_topic_rule(&self) -> EnableTopicRule<C, M, R>

pub fn get_behavior_model_training_summaries(
&self
) -> GetBehaviorModelTrainingSummaries<C, M, R>

pub fn get_buckets_aggregation(&self) -> GetBucketsAggregation<C, M, R>

pub fn get_cardinality(&self) -> GetCardinality<C, M, R>

pub fn get_effective_policies(&self) -> GetEffectivePolicies<C, M, R>

pub fn get_indexing_configuration(&self) -> GetIndexingConfiguration<C, M, R>

pub fn get_job_document(&self) -> GetJobDocument<C, M, R>

pub fn get_logging_options(&self) -> GetLoggingOptions<C, M, R>

pub fn get_ota_update(&self) -> GetOTAUpdate<C, M, R>

pub fn get_percentiles(&self) -> GetPercentiles<C, M, R>

pub fn get_policy(&self) -> GetPolicy<C, M, R>

pub fn get_policy_version(&self) -> GetPolicyVersion<C, M, R>

pub fn get_registration_code(&self) -> GetRegistrationCode<C, M, R>

pub fn get_statistics(&self) -> GetStatistics<C, M, R>

pub fn get_topic_rule(&self) -> GetTopicRule<C, M, R>

pub fn get_topic_rule_destination(&self) -> GetTopicRuleDestination<C, M, R>

pub fn get_v2_logging_options(&self) -> GetV2LoggingOptions<C, M, R>

pub fn list_active_violations(&self) -> ListActiveViolations<C, M, R>

pub fn list_attached_policies(&self) -> ListAttachedPolicies<C, M, R>

pub fn list_audit_findings(&self) -> ListAuditFindings<C, M, R>

pub fn list_audit_mitigation_actions_executions(
&self
) -> ListAuditMitigationActionsExecutions<C, M, R>

pub fn list_audit_mitigation_actions_tasks(
&self
) -> ListAuditMitigationActionsTasks<C, M, R>

pub fn list_audit_suppressions(&self) -> ListAuditSuppressions<C, M, R>

pub fn list_audit_tasks(&self) -> ListAuditTasks<C, M, R>

pub fn list_authorizers(&self) -> ListAuthorizers<C, M, R>

pub fn list_billing_groups(&self) -> ListBillingGroups<C, M, R>

pub fn list_ca_certificates(&self) -> ListCACertificates<C, M, R>

pub fn list_certificates(&self) -> ListCertificates<C, M, R>

pub fn list_certificates_by_ca(&self) -> ListCertificatesByCA<C, M, R>

pub fn list_custom_metrics(&self) -> ListCustomMetrics<C, M, R>

pub fn list_detect_mitigation_actions_executions(
&self
) -> ListDetectMitigationActionsExecutions<C, M, R>

pub fn list_detect_mitigation_actions_tasks(
&self
) -> ListDetectMitigationActionsTasks<C, M, R>

pub fn list_dimensions(&self) -> ListDimensions<C, M, R>

pub fn list_domain_configurations(&self) -> ListDomainConfigurations<C, M, R>

pub fn list_fleet_metrics(&self) -> ListFleetMetrics<C, M, R>

pub fn list_indices(&self) -> ListIndices<C, M, R>

pub fn list_job_executions_for_job(&self) -> ListJobExecutionsForJob<C, M, R>

pub fn list_job_executions_for_thing(
&self
) -> ListJobExecutionsForThing<C, M, R>

pub fn list_jobs(&self) -> ListJobs<C, M, R>

pub fn list_job_templates(&self) -> ListJobTemplates<C, M, R>

pub fn list_mitigation_actions(&self) -> ListMitigationActions<C, M, R>

pub fn list_ota_updates(&self) -> ListOTAUpdates<C, M, R>

pub fn list_outgoing_certificates(&self) -> ListOutgoingCertificates<C, M, R>

pub fn list_policies(&self) -> ListPolicies<C, M, R>

pub fn list_policy_principals(&self) -> ListPolicyPrincipals<C, M, R>

pub fn list_policy_versions(&self) -> ListPolicyVersions<C, M, R>

pub fn list_principal_policies(&self) -> ListPrincipalPolicies<C, M, R>

pub fn list_principal_things(&self) -> ListPrincipalThings<C, M, R>

pub fn list_provisioning_templates(&self) -> ListProvisioningTemplates<C, M, R>

pub fn list_provisioning_template_versions(
&self
) -> ListProvisioningTemplateVersions<C, M, R>

pub fn list_role_aliases(&self) -> ListRoleAliases<C, M, R>

pub fn list_scheduled_audits(&self) -> ListScheduledAudits<C, M, R>

pub fn list_security_profiles(&self) -> ListSecurityProfiles<C, M, R>

pub fn list_security_profiles_for_target(
&self
) -> ListSecurityProfilesForTarget<C, M, R>

pub fn list_streams(&self) -> ListStreams<C, M, R>

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

pub fn list_targets_for_policy(&self) -> ListTargetsForPolicy<C, M, R>

pub fn list_targets_for_security_profile(
&self
) -> ListTargetsForSecurityProfile<C, M, R>

pub fn list_thing_groups(&self) -> ListThingGroups<C, M, R>

pub fn list_thing_groups_for_thing(&self) -> ListThingGroupsForThing<C, M, R>

pub fn list_thing_principals(&self) -> ListThingPrincipals<C, M, R>

pub fn list_thing_registration_task_reports(
&self
) -> ListThingRegistrationTaskReports<C, M, R>

pub fn list_thing_registration_tasks(
&self
) -> ListThingRegistrationTasks<C, M, R>

pub fn list_things(&self) -> ListThings<C, M, R>

pub fn list_things_in_billing_group(&self) -> ListThingsInBillingGroup<C, M, R>

pub fn list_things_in_thing_group(&self) -> ListThingsInThingGroup<C, M, R>

pub fn list_thing_types(&self) -> ListThingTypes<C, M, R>

pub fn list_topic_rule_destinations(&self) -> ListTopicRuleDestinations<C, M, R>

pub fn list_topic_rules(&self) -> ListTopicRules<C, M, R>

pub fn list_v2_logging_levels(&self) -> ListV2LoggingLevels<C, M, R>

pub fn list_violation_events(&self) -> ListViolationEvents<C, M, R>

pub fn register_ca_certificate(&self) -> RegisterCACertificate<C, M, R>

pub fn register_certificate(&self) -> RegisterCertificate<C, M, R>

pub fn register_certificate_without_ca(
&self
) -> RegisterCertificateWithoutCA<C, M, R>

pub fn register_thing(&self) -> RegisterThing<C, M, R>

pub fn reject_certificate_transfer(&self) -> RejectCertificateTransfer<C, M, R>

pub fn remove_thing_from_billing_group(
&self
) -> RemoveThingFromBillingGroup<C, M, R>

pub fn remove_thing_from_thing_group(
&self
) -> RemoveThingFromThingGroup<C, M, R>

pub fn replace_topic_rule(&self) -> ReplaceTopicRule<C, M, R>

pub fn search_index(&self) -> SearchIndex<C, M, R>

pub fn set_default_authorizer(&self) -> SetDefaultAuthorizer<C, M, R>

pub fn set_default_policy_version(&self) -> SetDefaultPolicyVersion<C, M, R>

pub fn set_logging_options(&self) -> SetLoggingOptions<C, M, R>

pub fn set_v2_logging_level(&self) -> SetV2LoggingLevel<C, M, R>

pub fn set_v2_logging_options(&self) -> SetV2LoggingOptions<C, M, R>

pub fn start_audit_mitigation_actions_task(
&self
) -> StartAuditMitigationActionsTask<C, M, R>

pub fn start_detect_mitigation_actions_task(
&self
) -> StartDetectMitigationActionsTask<C, M, R>

pub fn start_on_demand_audit_task(&self) -> StartOnDemandAuditTask<C, M, R>

pub fn start_thing_registration_task(
&self
) -> StartThingRegistrationTask<C, M, R>

pub fn stop_thing_registration_task(&self) -> StopThingRegistrationTask<C, M, R>

pub fn tag_resource(&self) -> TagResource<C, M, R>

pub fn test_authorization(&self) -> TestAuthorization<C, M, R>

pub fn test_invoke_authorizer(&self) -> TestInvokeAuthorizer<C, M, R>

pub fn transfer_certificate(&self) -> TransferCertificate<C, M, R>

pub fn untag_resource(&self) -> UntagResource<C, M, R>

pub fn update_account_audit_configuration(
&self
) -> UpdateAccountAuditConfiguration<C, M, R>

pub fn update_audit_suppression(&self) -> UpdateAuditSuppression<C, M, R>

pub fn update_authorizer(&self) -> UpdateAuthorizer<C, M, R>

pub fn update_billing_group(&self) -> UpdateBillingGroup<C, M, R>

pub fn update_ca_certificate(&self) -> UpdateCACertificate<C, M, R>

pub fn update_certificate(&self) -> UpdateCertificate<C, M, R>

pub fn update_custom_metric(&self) -> UpdateCustomMetric<C, M, R>

pub fn update_dimension(&self) -> UpdateDimension<C, M, R>

pub fn update_domain_configuration(&self) -> UpdateDomainConfiguration<C, M, R>

pub fn update_dynamic_thing_group(&self) -> UpdateDynamicThingGroup<C, M, R>

pub fn update_event_configurations(&self) -> UpdateEventConfigurations<C, M, R>

pub fn update_fleet_metric(&self) -> UpdateFleetMetric<C, M, R>

pub fn update_indexing_configuration(
&self
) -> UpdateIndexingConfiguration<C, M, R>

pub fn update_job(&self) -> UpdateJob<C, M, R>

pub fn update_mitigation_action(&self) -> UpdateMitigationAction<C, M, R>

pub fn update_provisioning_template(
&self
) -> UpdateProvisioningTemplate<C, M, R>

pub fn update_role_alias(&self) -> UpdateRoleAlias<C, M, R>

pub fn update_scheduled_audit(&self) -> UpdateScheduledAudit<C, M, R>

pub fn update_security_profile(&self) -> UpdateSecurityProfile<C, M, R>

pub fn update_stream(&self) -> UpdateStream<C, M, R>

pub fn update_thing(&self) -> UpdateThing<C, M, R>

pub fn update_thing_group(&self) -> UpdateThingGroup<C, M, R>

pub fn update_thing_groups_for_thing(
&self
) -> UpdateThingGroupsForThing<C, M, R>

pub fn update_topic_rule_destination(
&self
) -> UpdateTopicRuleDestination<C, M, R>

pub fn validate_security_profile_behaviors(
&self
) -> ValidateSecurityProfileBehaviors<C, M, R>

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

pub fn from_conf(conf: Config) -> Self