Module aws_sdk_iot::error

Modules

certificate_conflict_exception

See CertificateConflictException

certificate_state_exception

See CertificateStateException

certificate_validation_exception

See CertificateValidationException

conflict_exception

See ConflictException

conflicting_resource_update_exception

See ConflictingResourceUpdateException

delete_conflict_exception

See DeleteConflictException

index_not_ready_exception

See IndexNotReadyException

internal_exception

See InternalException

internal_failure_exception

See InternalFailureException

invalid_aggregation_exception

See InvalidAggregationException

invalid_query_exception

See InvalidQueryException

invalid_request_exception

See InvalidRequestException

invalid_response_exception

See InvalidResponseException

invalid_state_transition_exception

See InvalidStateTransitionException

limit_exceeded_exception

See LimitExceededException

malformed_policy_exception

See MalformedPolicyException

not_configured_exception

See NotConfiguredException

registration_code_validation_exception

See RegistrationCodeValidationException

resource_already_exists_exception

See ResourceAlreadyExistsException

resource_not_found_exception

See ResourceNotFoundException

resource_registration_failure_exception

See ResourceRegistrationFailureException

service_unavailable_exception

See ServiceUnavailableException

sql_parse_exception

See SqlParseException

task_already_exists_exception

See TaskAlreadyExistsException

throttling_exception

See ThrottlingException

transfer_already_completed_exception

See TransferAlreadyCompletedException

transfer_conflict_exception

See TransferConflictException

unauthorized_exception

See UnauthorizedException

version_conflict_exception

See VersionConflictException

versions_limit_exceeded_exception

See VersionsLimitExceededException

Structs

AcceptCertificateTransferError

AddThingToBillingGroupError

AddThingToThingGroupError

AssociateTargetsWithJobError

AttachPolicyError

AttachPrincipalPolicyError

AttachSecurityProfileError

AttachThingPrincipalError

CancelAuditMitigationActionsTaskError

CancelAuditTaskError

CancelCertificateTransferError

CancelDetectMitigationActionsTaskError

CancelJobError

CancelJobExecutionError

CertificateConflictException

Unable to verify the CA certificate used to sign the device certificate you are

CertificateStateException

The certificate operation is not allowed.

CertificateValidationException

The certificate is invalid.

ClearDefaultAuthorizerError

ConfirmTopicRuleDestinationError

ConflictException

A resource with the same name already exists.

ConflictingResourceUpdateException

A conflicting resource update exception. This exception is thrown when two pending

CreateAuditSuppressionError

CreateAuthorizerError

CreateBillingGroupError

CreateCertificateFromCsrError

CreateCustomMetricError

CreateDimensionError

CreateDomainConfigurationError

CreateDynamicThingGroupError

CreateFleetMetricError

CreateJobError

CreateJobTemplateError

CreateKeysAndCertificateError

CreateMitigationActionError

CreateOTAUpdateError

CreatePolicyError

CreatePolicyVersionError

CreateProvisioningClaimError

CreateProvisioningTemplateError

CreateProvisioningTemplateVersionError

CreateRoleAliasError

CreateScheduledAuditError

CreateSecurityProfileError

CreateStreamError

CreateThingError

CreateThingGroupError

CreateThingTypeError

CreateTopicRuleDestinationError

CreateTopicRuleError

DeleteAccountAuditConfigurationError

DeleteAuditSuppressionError

DeleteAuthorizerError

DeleteBillingGroupError

DeleteCACertificateError

DeleteCertificateError

DeleteConflictException

You can't delete the resource because it is attached to one or more

DeleteCustomMetricError

DeleteDimensionError

DeleteDomainConfigurationError

DeleteDynamicThingGroupError

DeleteFleetMetricError

DeleteJobError

DeleteJobExecutionError

DeleteJobTemplateError

DeleteMitigationActionError

DeleteOTAUpdateError

DeletePolicyError

DeletePolicyVersionError

DeleteProvisioningTemplateError

DeleteProvisioningTemplateVersionError

DeleteRegistrationCodeError

DeleteRoleAliasError

DeleteScheduledAuditError

DeleteSecurityProfileError

DeleteStreamError

DeleteThingError

DeleteThingGroupError

DeleteThingTypeError

DeleteTopicRuleDestinationError

DeleteTopicRuleError

DeleteV2LoggingLevelError

DeprecateThingTypeError

DescribeAccountAuditConfigurationError

DescribeAuditFindingError

DescribeAuditMitigationActionsTaskError

DescribeAuditSuppressionError

DescribeAuditTaskError

DescribeAuthorizerError

DescribeBillingGroupError

DescribeCACertificateError

DescribeCertificateError

DescribeCustomMetricError

DescribeDefaultAuthorizerError

DescribeDetectMitigationActionsTaskError

DescribeDimensionError

DescribeDomainConfigurationError

DescribeEndpointError

DescribeEventConfigurationsError

DescribeFleetMetricError

DescribeIndexError

DescribeJobError

DescribeJobExecutionError

DescribeJobTemplateError

DescribeMitigationActionError

DescribeProvisioningTemplateError

DescribeProvisioningTemplateVersionError

DescribeRoleAliasError

DescribeScheduledAuditError

DescribeSecurityProfileError

DescribeStreamError

DescribeThingError

DescribeThingGroupError

DescribeThingRegistrationTaskError

DescribeThingTypeError

DetachPolicyError

DetachPrincipalPolicyError

DetachSecurityProfileError

DetachThingPrincipalError

DisableTopicRuleError

EnableTopicRuleError

GetBehaviorModelTrainingSummariesError

GetBucketsAggregationError

GetCardinalityError

GetEffectivePoliciesError

GetIndexingConfigurationError

GetJobDocumentError

GetLoggingOptionsError

GetOTAUpdateError

GetPercentilesError

GetPolicyError

GetPolicyVersionError

GetRegistrationCodeError

GetStatisticsError

GetTopicRuleDestinationError

GetTopicRuleError

GetV2LoggingOptionsError

IndexNotReadyException

The index is not ready.

InternalException

An unexpected error has occurred.

InternalFailureException

An unexpected error has occurred.

InvalidAggregationException

The aggregation is invalid.

InvalidQueryException

The query is invalid.

InvalidRequestException

The request is not valid.

InvalidResponseException

The response is invalid.

InvalidStateTransitionException

An attempt was made to change to an invalid state, for example by deleting a job or a

LimitExceededException

A limit has been exceeded.

ListActiveViolationsError

ListAttachedPoliciesError

ListAuditFindingsError

ListAuditMitigationActionsExecutionsError

ListAuditMitigationActionsTasksError

ListAuditSuppressionsError

ListAuditTasksError

ListAuthorizersError

ListBillingGroupsError

ListCACertificatesError

ListCertificatesByCAError

ListCertificatesError

ListCustomMetricsError

ListDetectMitigationActionsExecutionsError

ListDetectMitigationActionsTasksError

ListDimensionsError

ListDomainConfigurationsError

ListFleetMetricsError

ListIndicesError

ListJobExecutionsForJobError

ListJobExecutionsForThingError

ListJobTemplatesError

ListJobsError

ListMitigationActionsError

ListOTAUpdatesError

ListOutgoingCertificatesError

ListPoliciesError

ListPolicyPrincipalsError

ListPolicyVersionsError

ListPrincipalPoliciesError

ListPrincipalThingsError

ListProvisioningTemplateVersionsError

ListProvisioningTemplatesError

ListRoleAliasesError

ListScheduledAuditsError

ListSecurityProfilesError

ListSecurityProfilesForTargetError

ListStreamsError

ListTagsForResourceError

ListTargetsForPolicyError

ListTargetsForSecurityProfileError

ListThingGroupsError

ListThingGroupsForThingError

ListThingPrincipalsError

ListThingRegistrationTaskReportsError

ListThingRegistrationTasksError

ListThingTypesError

ListThingsError

ListThingsInBillingGroupError

ListThingsInThingGroupError

ListTopicRuleDestinationsError

ListTopicRulesError

ListV2LoggingLevelsError

ListViolationEventsError

MalformedPolicyException

The policy documentation is not valid.

NotConfiguredException

The resource is not configured.

RegisterCACertificateError

RegisterCertificateError

RegisterCertificateWithoutCAError

RegisterThingError

RegistrationCodeValidationException

The registration code is invalid.

RejectCertificateTransferError

RemoveThingFromBillingGroupError

RemoveThingFromThingGroupError

ReplaceTopicRuleError

ResourceAlreadyExistsException

The resource already exists.

ResourceNotFoundException

The specified resource does not exist.

ResourceRegistrationFailureException

The resource registration failed.

SearchIndexError

ServiceUnavailableException

The service is temporarily unavailable.

SetDefaultAuthorizerError

SetDefaultPolicyVersionError

SetLoggingOptionsError

SetV2LoggingLevelError

SetV2LoggingOptionsError

SqlParseException

The Rule-SQL expression can't be parsed correctly.

StartAuditMitigationActionsTaskError

StartDetectMitigationActionsTaskError

StartOnDemandAuditTaskError

StartThingRegistrationTaskError

StopThingRegistrationTaskError

TagResourceError

TaskAlreadyExistsException

TestAuthorizationError

TestInvokeAuthorizerError

ThrottlingException

The rate exceeds the limit.

TransferAlreadyCompletedException

You can't revert the certificate transfer because the transfer is already

TransferCertificateError

TransferConflictException

You can't transfer the certificate because authorization policies are still

UnauthorizedException

You are not authorized to perform this operation.

UntagResourceError

UpdateAccountAuditConfigurationError

UpdateAuditSuppressionError

UpdateAuthorizerError

UpdateBillingGroupError

UpdateCACertificateError

UpdateCertificateError

UpdateCustomMetricError

UpdateDimensionError

UpdateDomainConfigurationError

UpdateDynamicThingGroupError

UpdateEventConfigurationsError

UpdateFleetMetricError

UpdateIndexingConfigurationError

UpdateJobError

UpdateMitigationActionError

UpdateProvisioningTemplateError

UpdateRoleAliasError

UpdateScheduledAuditError

UpdateSecurityProfileError

UpdateStreamError

UpdateThingError

UpdateThingGroupError

UpdateThingGroupsForThingError

UpdateTopicRuleDestinationError

ValidateSecurityProfileBehaviorsError

VersionConflictException

An exception thrown when the version of an entity specified with the

VersionsLimitExceededException

The number of policy versions exceeds the limit.

Enums

AcceptCertificateTransferErrorKind

AddThingToBillingGroupErrorKind

AddThingToThingGroupErrorKind

AssociateTargetsWithJobErrorKind

AttachPolicyErrorKind

AttachPrincipalPolicyErrorKind

AttachSecurityProfileErrorKind

AttachThingPrincipalErrorKind

CancelAuditMitigationActionsTaskErrorKind

CancelAuditTaskErrorKind

CancelCertificateTransferErrorKind

CancelDetectMitigationActionsTaskErrorKind

CancelJobErrorKind

CancelJobExecutionErrorKind

ClearDefaultAuthorizerErrorKind

ConfirmTopicRuleDestinationErrorKind

CreateAuditSuppressionErrorKind

CreateAuthorizerErrorKind

CreateBillingGroupErrorKind

CreateCertificateFromCsrErrorKind

CreateCustomMetricErrorKind

CreateDimensionErrorKind

CreateDomainConfigurationErrorKind

CreateDynamicThingGroupErrorKind

CreateFleetMetricErrorKind

CreateJobErrorKind

CreateJobTemplateErrorKind

CreateKeysAndCertificateErrorKind

CreateMitigationActionErrorKind

CreateOTAUpdateErrorKind

CreatePolicyErrorKind

CreatePolicyVersionErrorKind

CreateProvisioningClaimErrorKind

CreateProvisioningTemplateErrorKind

CreateProvisioningTemplateVersionErrorKind

CreateRoleAliasErrorKind

CreateScheduledAuditErrorKind

CreateSecurityProfileErrorKind

CreateStreamErrorKind

CreateThingErrorKind

CreateThingGroupErrorKind

CreateThingTypeErrorKind

CreateTopicRuleDestinationErrorKind

CreateTopicRuleErrorKind

DeleteAccountAuditConfigurationErrorKind

DeleteAuditSuppressionErrorKind

DeleteAuthorizerErrorKind

DeleteBillingGroupErrorKind

DeleteCACertificateErrorKind

DeleteCertificateErrorKind

DeleteCustomMetricErrorKind

DeleteDimensionErrorKind

DeleteDomainConfigurationErrorKind

DeleteDynamicThingGroupErrorKind

DeleteFleetMetricErrorKind

DeleteJobErrorKind

DeleteJobExecutionErrorKind

DeleteJobTemplateErrorKind

DeleteMitigationActionErrorKind

DeleteOTAUpdateErrorKind

DeletePolicyErrorKind

DeletePolicyVersionErrorKind

DeleteProvisioningTemplateErrorKind

DeleteProvisioningTemplateVersionErrorKind

DeleteRegistrationCodeErrorKind

DeleteRoleAliasErrorKind

DeleteScheduledAuditErrorKind

DeleteSecurityProfileErrorKind

DeleteStreamErrorKind

DeleteThingErrorKind

DeleteThingGroupErrorKind

DeleteThingTypeErrorKind

DeleteTopicRuleDestinationErrorKind

DeleteTopicRuleErrorKind

DeleteV2LoggingLevelErrorKind

DeprecateThingTypeErrorKind

DescribeAccountAuditConfigurationErrorKind

DescribeAuditFindingErrorKind

DescribeAuditMitigationActionsTaskErrorKind

DescribeAuditSuppressionErrorKind

DescribeAuditTaskErrorKind

DescribeAuthorizerErrorKind

DescribeBillingGroupErrorKind

DescribeCACertificateErrorKind

DescribeCertificateErrorKind

DescribeCustomMetricErrorKind

DescribeDefaultAuthorizerErrorKind

DescribeDetectMitigationActionsTaskErrorKind

DescribeDimensionErrorKind

DescribeDomainConfigurationErrorKind

DescribeEndpointErrorKind

DescribeEventConfigurationsErrorKind

DescribeFleetMetricErrorKind

DescribeIndexErrorKind

DescribeJobErrorKind

DescribeJobExecutionErrorKind

DescribeJobTemplateErrorKind

DescribeMitigationActionErrorKind

DescribeProvisioningTemplateErrorKind

DescribeProvisioningTemplateVersionErrorKind

DescribeRoleAliasErrorKind

DescribeScheduledAuditErrorKind

DescribeSecurityProfileErrorKind

DescribeStreamErrorKind

DescribeThingErrorKind

DescribeThingGroupErrorKind

DescribeThingRegistrationTaskErrorKind

DescribeThingTypeErrorKind

DetachPolicyErrorKind

DetachPrincipalPolicyErrorKind

DetachSecurityProfileErrorKind

DetachThingPrincipalErrorKind

DisableTopicRuleErrorKind

EnableTopicRuleErrorKind

GetBehaviorModelTrainingSummariesErrorKind

GetBucketsAggregationErrorKind

GetCardinalityErrorKind

GetEffectivePoliciesErrorKind

GetIndexingConfigurationErrorKind

GetJobDocumentErrorKind

GetLoggingOptionsErrorKind

GetOTAUpdateErrorKind

GetPercentilesErrorKind

GetPolicyErrorKind

GetPolicyVersionErrorKind

GetRegistrationCodeErrorKind

GetStatisticsErrorKind

GetTopicRuleDestinationErrorKind

GetTopicRuleErrorKind

GetV2LoggingOptionsErrorKind

ListActiveViolationsErrorKind

ListAttachedPoliciesErrorKind

ListAuditFindingsErrorKind

ListAuditMitigationActionsExecutionsErrorKind

ListAuditMitigationActionsTasksErrorKind

ListAuditSuppressionsErrorKind

ListAuditTasksErrorKind

ListAuthorizersErrorKind

ListBillingGroupsErrorKind

ListCACertificatesErrorKind

ListCertificatesByCAErrorKind

ListCertificatesErrorKind

ListCustomMetricsErrorKind

ListDetectMitigationActionsExecutionsErrorKind

ListDetectMitigationActionsTasksErrorKind

ListDimensionsErrorKind

ListDomainConfigurationsErrorKind

ListFleetMetricsErrorKind

ListIndicesErrorKind

ListJobExecutionsForJobErrorKind

ListJobExecutionsForThingErrorKind

ListJobTemplatesErrorKind

ListJobsErrorKind

ListMitigationActionsErrorKind

ListOTAUpdatesErrorKind

ListOutgoingCertificatesErrorKind

ListPoliciesErrorKind

ListPolicyPrincipalsErrorKind

ListPolicyVersionsErrorKind

ListPrincipalPoliciesErrorKind

ListPrincipalThingsErrorKind

ListProvisioningTemplateVersionsErrorKind

ListProvisioningTemplatesErrorKind

ListRoleAliasesErrorKind

ListScheduledAuditsErrorKind

ListSecurityProfilesErrorKind

ListSecurityProfilesForTargetErrorKind

ListStreamsErrorKind

ListTagsForResourceErrorKind

ListTargetsForPolicyErrorKind

ListTargetsForSecurityProfileErrorKind

ListThingGroupsErrorKind

ListThingGroupsForThingErrorKind

ListThingPrincipalsErrorKind

ListThingRegistrationTaskReportsErrorKind

ListThingRegistrationTasksErrorKind

ListThingTypesErrorKind

ListThingsErrorKind

ListThingsInBillingGroupErrorKind

ListThingsInThingGroupErrorKind

ListTopicRuleDestinationsErrorKind

ListTopicRulesErrorKind

ListV2LoggingLevelsErrorKind

ListViolationEventsErrorKind

RegisterCACertificateErrorKind

RegisterCertificateErrorKind

RegisterCertificateWithoutCAErrorKind

RegisterThingErrorKind

RejectCertificateTransferErrorKind

RemoveThingFromBillingGroupErrorKind

RemoveThingFromThingGroupErrorKind

ReplaceTopicRuleErrorKind

SearchIndexErrorKind

SetDefaultAuthorizerErrorKind

SetDefaultPolicyVersionErrorKind

SetLoggingOptionsErrorKind

SetV2LoggingLevelErrorKind

SetV2LoggingOptionsErrorKind

StartAuditMitigationActionsTaskErrorKind

StartDetectMitigationActionsTaskErrorKind

StartOnDemandAuditTaskErrorKind

StartThingRegistrationTaskErrorKind

StopThingRegistrationTaskErrorKind

TagResourceErrorKind

TestAuthorizationErrorKind

TestInvokeAuthorizerErrorKind

TransferCertificateErrorKind

UntagResourceErrorKind

UpdateAccountAuditConfigurationErrorKind

UpdateAuditSuppressionErrorKind

UpdateAuthorizerErrorKind

UpdateBillingGroupErrorKind

UpdateCACertificateErrorKind

UpdateCertificateErrorKind

UpdateCustomMetricErrorKind

UpdateDimensionErrorKind

UpdateDomainConfigurationErrorKind

UpdateDynamicThingGroupErrorKind

UpdateEventConfigurationsErrorKind

UpdateFleetMetricErrorKind

UpdateIndexingConfigurationErrorKind

UpdateJobErrorKind

UpdateMitigationActionErrorKind

UpdateProvisioningTemplateErrorKind

UpdateRoleAliasErrorKind

UpdateScheduledAuditErrorKind

UpdateSecurityProfileErrorKind

UpdateStreamErrorKind

UpdateThingErrorKind

UpdateThingGroupErrorKind

UpdateThingGroupsForThingErrorKind

UpdateTopicRuleDestinationErrorKind

ValidateSecurityProfileBehaviorsErrorKind