Module aws_sdk_iot::model

Modules

abort_config

See AbortConfig

abort_criteria

See AbortCriteria

action

See Action

active_violation

See ActiveViolation

add_things_to_thing_group_params

See AddThingsToThingGroupParams

aggregation_type

See AggregationType

alert_target

See AlertTarget

allowed

See Allowed

asset_property_timestamp

See AssetPropertyTimestamp

asset_property_value

See AssetPropertyValue

attribute_payload

See AttributePayload

audit_check_configuration

See AuditCheckConfiguration

audit_check_details

See AuditCheckDetails

audit_finding

See AuditFinding

audit_mitigation_action_execution_metadata

See AuditMitigationActionExecutionMetadata

audit_mitigation_actions_task_metadata

See AuditMitigationActionsTaskMetadata

audit_mitigation_actions_task_target

See AuditMitigationActionsTaskTarget

audit_notification_target

See AuditNotificationTarget

audit_suppression

See AuditSuppression

audit_task_metadata

See AuditTaskMetadata

auth_info

See AuthInfo

auth_result

See AuthResult

authorizer_config

See AuthorizerConfig

authorizer_description

See AuthorizerDescription

authorizer_summary

See AuthorizerSummary

aws_job_abort_config

See AwsJobAbortConfig

aws_job_abort_criteria

See AwsJobAbortCriteria

aws_job_executions_rollout_config

See AwsJobExecutionsRolloutConfig

aws_job_exponential_rollout_rate

See AwsJobExponentialRolloutRate

aws_job_presigned_url_config

See AwsJobPresignedUrlConfig

aws_job_rate_increase_criteria

See AwsJobRateIncreaseCriteria

aws_job_timeout_config

See AwsJobTimeoutConfig

behavior

See Behavior

behavior_criteria

See BehaviorCriteria

behavior_model_training_summary

See BehaviorModelTrainingSummary

billing_group_metadata

See BillingGroupMetadata

billing_group_properties

See BillingGroupProperties

bucket

See Bucket

buckets_aggregation_type

See BucketsAggregationType

ca_certificate

See CaCertificate

ca_certificate_description

See CaCertificateDescription

certificate

See Certificate

certificate_description

See CertificateDescription

certificate_validity

See CertificateValidity

cloudwatch_alarm_action

See CloudwatchAlarmAction

cloudwatch_logs_action

See CloudwatchLogsAction

cloudwatch_metric_action

See CloudwatchMetricAction

code_signing

See CodeSigning

code_signing_certificate_chain

See CodeSigningCertificateChain

code_signing_signature

See CodeSigningSignature

configuration

See Configuration

custom_code_signing

See CustomCodeSigning

denied

See Denied

destination

See Destination

detect_mitigation_action_execution

See DetectMitigationActionExecution

detect_mitigation_actions_task_statistics

See DetectMitigationActionsTaskStatistics

detect_mitigation_actions_task_summary

See DetectMitigationActionsTaskSummary

detect_mitigation_actions_task_target

See DetectMitigationActionsTaskTarget

domain_configuration_summary

See DomainConfigurationSummary

dynamo_d_bv2_action

See DynamoDBv2Action

dynamo_db_action

See DynamoDbAction

effective_policy

See EffectivePolicy

elasticsearch_action

See ElasticsearchAction

enable_io_t_logging_params

See EnableIoTLoggingParams

error_info

See ErrorInfo

explicit_deny

See ExplicitDeny

exponential_rollout_rate

See ExponentialRolloutRate

field

See Field

file_location

See FileLocation

firehose_action

See FirehoseAction

fleet_metric_name_and_arn

See FleetMetricNameAndArn

group_name_and_arn

See GroupNameAndArn

http_action

See HttpAction

http_action_header

See HttpActionHeader

http_authorization

See HttpAuthorization

http_context

See HttpContext

http_url_destination_configuration

See HttpUrlDestinationConfiguration

http_url_destination_properties

See HttpUrlDestinationProperties

http_url_destination_summary

See HttpUrlDestinationSummary

implicit_deny

See ImplicitDeny

iot_analytics_action

See IotAnalyticsAction

iot_events_action

See IotEventsAction

iot_site_wise_action

See IotSiteWiseAction

job

See Job

job_execution

See JobExecution

job_execution_status_details

See JobExecutionStatusDetails

job_execution_summary

See JobExecutionSummary

job_execution_summary_for_job

See JobExecutionSummaryForJob

job_execution_summary_for_thing

See JobExecutionSummaryForThing

job_executions_rollout_config

See JobExecutionsRolloutConfig

job_process_details

See JobProcessDetails

job_summary

See JobSummary

job_template_summary

See JobTemplateSummary

kafka_action

See KafkaAction

key_pair

See KeyPair

kinesis_action

See KinesisAction

lambda_action

See LambdaAction

log_target

See LogTarget

log_target_configuration

See LogTargetConfiguration

logging_options_payload

See LoggingOptionsPayload

machine_learning_detection_config

See MachineLearningDetectionConfig

metric_dimension

See MetricDimension

metric_to_retain

See MetricToRetain

metric_value

See MetricValue

mitigation_action

See MitigationAction

mitigation_action_identifier

See MitigationActionIdentifier

mitigation_action_params

See MitigationActionParams

mqtt_context

See MqttContext

non_compliant_resource

See NonCompliantResource

open_search_action

See OpenSearchAction

ota_update_file

See OtaUpdateFile

ota_update_info

See OtaUpdateInfo

ota_update_summary

See OtaUpdateSummary

outgoing_certificate

See OutgoingCertificate

percent_pair

See PercentPair

policy

See Policy

policy_version

See PolicyVersion

policy_version_identifier

See PolicyVersionIdentifier

presigned_url_config

See PresignedUrlConfig

provisioning_hook

See ProvisioningHook

provisioning_template_summary

See ProvisioningTemplateSummary

provisioning_template_version_summary

See ProvisioningTemplateVersionSummary

publish_finding_to_sns_params

See PublishFindingToSnsParams

put_asset_property_value_entry

See PutAssetPropertyValueEntry

put_item_input

See PutItemInput

rate_increase_criteria

See RateIncreaseCriteria

registration_config

See RegistrationConfig

related_resource

See RelatedResource

replace_default_policy_version_params

See ReplaceDefaultPolicyVersionParams

republish_action

See RepublishAction

resource_identifier

See ResourceIdentifier

role_alias_description

See RoleAliasDescription

s3_action

See S3Action

s3_destination

See S3Destination

s3_location

See S3Location

salesforce_action

See SalesforceAction

scheduled_audit_metadata

See ScheduledAuditMetadata

security_profile_identifier

See SecurityProfileIdentifier

security_profile_target

See SecurityProfileTarget

security_profile_target_mapping

See SecurityProfileTargetMapping

server_certificate_summary

See ServerCertificateSummary

sig_v4_authorization

See SigV4Authorization

signing_profile_parameter

See SigningProfileParameter

sns_action

See SnsAction

sqs_action

See SqsAction

start_signing_job_parameter

See StartSigningJobParameter

statistical_threshold

See StatisticalThreshold

statistics

See Statistics

step_functions_action

See StepFunctionsAction

stream

See Stream

stream_file

See StreamFile

stream_info

See StreamInfo

stream_summary

See StreamSummary

tag

See Tag

task_statistics

See TaskStatistics

task_statistics_for_audit_check

See TaskStatisticsForAuditCheck

terms_aggregation

See TermsAggregation

thing_attribute

See ThingAttribute

thing_connectivity

See ThingConnectivity

thing_document

See ThingDocument

thing_group_document

See ThingGroupDocument

thing_group_indexing_configuration

See ThingGroupIndexingConfiguration

thing_group_metadata

See ThingGroupMetadata

thing_group_properties

See ThingGroupProperties

thing_indexing_configuration

See ThingIndexingConfiguration

thing_type_definition

See ThingTypeDefinition

thing_type_metadata

See ThingTypeMetadata

thing_type_properties

See ThingTypeProperties

timeout_config

See TimeoutConfig

timestream_action

See TimestreamAction

timestream_dimension

See TimestreamDimension

timestream_timestamp

See TimestreamTimestamp

tls_context

See TlsContext

topic_rule

See TopicRule

topic_rule_destination

See TopicRuleDestination

topic_rule_destination_configuration

See TopicRuleDestinationConfiguration

topic_rule_destination_summary

See TopicRuleDestinationSummary

topic_rule_list_item

See TopicRuleListItem

topic_rule_payload

See TopicRulePayload

transfer_data

See TransferData

update_ca_certificate_params

See UpdateCaCertificateParams

update_device_certificate_params

See UpdateDeviceCertificateParams

validation_error

See ValidationError

violation_event

See ViolationEvent

violation_event_additional_info

See ViolationEventAdditionalInfo

violation_event_occurrence_range

See ViolationEventOccurrenceRange

vpc_destination_configuration

See VpcDestinationConfiguration

vpc_destination_properties

See VpcDestinationProperties

vpc_destination_summary

See VpcDestinationSummary

Structs

AbortConfig

The criteria that determine when and how a job abort takes place.

AbortCriteria

The criteria that determine when and how a job abort takes place.

Action

Describes the actions associated with a rule.

ActiveViolation

Information about an active Device Defender security profile behavior violation.

AddThingsToThingGroupParams

Parameters used when defining a mitigation action that move a set of things to a thing group.

AggregationType

The type of aggregation queries.

AlertTarget

A structure containing the alert target ARN and the role ARN.

Allowed

Contains information that allowed the authorization.

AssetPropertyTimestamp

An asset property timestamp entry containing the following information.

AssetPropertyValue

An asset property value entry containing the following information.

AttributePayload

The attribute payload.

AuditCheckConfiguration

Which audit checks are enabled and disabled for this account.

AuditCheckDetails

Information about the audit check.

AuditFinding

The findings (results) of the audit.

AuditMitigationActionExecutionMetadata

Returned by ListAuditMitigationActionsTask, this object contains information that describes a mitigation action that has been started.

AuditMitigationActionsTaskMetadata

Information about an audit mitigation actions task that is returned by ListAuditMitigationActionsTasks.

AuditMitigationActionsTaskTarget

Used in MitigationActionParams, this information identifies the target findings to which the mitigation actions are applied. Only one entry appears.

AuditNotificationTarget

Information about the targets to which audit notifications are sent.

AuditSuppression

Filters out specific findings of a Device Defender audit.

AuditTaskMetadata

The audits that were performed.

AuthInfo

A collection of authorization information.

AuthResult

The authorizer result.

AuthorizerConfig

An object that specifies the authorization service for a domain.

AuthorizerDescription

The authorizer description.

AuthorizerSummary

The authorizer summary.

AwsJobAbortConfig

The criteria that determine when and how a job abort takes place.

AwsJobAbortCriteria

The criteria that determine when and how a job abort takes place.

AwsJobExecutionsRolloutConfig

Configuration for the rollout of OTA updates.

AwsJobExponentialRolloutRate

The rate of increase for a job rollout. This parameter allows you to define an exponential rate

AwsJobPresignedUrlConfig

Configuration information for pre-signed URLs. Valid when protocols

AwsJobRateIncreaseCriteria

The criteria to initiate the increase in rate of rollout for a job.

AwsJobTimeoutConfig

Specifies the amount of time each device has to finish its execution of the job. A timer is

Behavior

A Device Defender security profile behavior.

BehaviorCriteria

The criteria by which the behavior is determined to be normal.

BehaviorModelTrainingSummary

BillingGroupMetadata

Additional information about the billing group.

BillingGroupProperties

The properties of a billing group.

Bucket

A count of documents that meets a specific aggregation criteria.

BucketsAggregationType

The type of bucketed aggregation performed.

CaCertificate

A CA certificate.

CaCertificateDescription

Describes a CA certificate.

Certificate

Information about a certificate.

CertificateDescription

Describes a certificate.

CertificateValidity

When the certificate is valid.

CloudwatchAlarmAction

Describes an action that updates a CloudWatch alarm.

CloudwatchLogsAction

Describes an action that sends data to CloudWatch Logs.

CloudwatchMetricAction

Describes an action that captures a CloudWatch metric.

CodeSigning

Describes the method to use when code signing a file.

CodeSigningCertificateChain

Describes the certificate chain being used when code signing a file.

CodeSigningSignature

Describes the signature for a file.

Configuration

Configuration.

CustomCodeSigning

Describes a custom method used to code sign a file.

Denied

Contains information that denied the authorization.

Destination

Describes the location of the updated firmware.

DetectMitigationActionExecution

DetectMitigationActionsTaskStatistics

DetectMitigationActionsTaskSummary

DetectMitigationActionsTaskTarget

DomainConfigurationSummary

The summary of a domain configuration. A domain configuration specifies custom IoT-specific information about a domain.

DynamoDBv2Action

Describes an action to write to a DynamoDB table.

DynamoDbAction

Describes an action to write to a DynamoDB table.

EffectivePolicy

The policy that has the effect on the authorization results.

ElasticsearchAction

Describes an action that writes data to an Amazon Elasticsearch Service

EnableIoTLoggingParams

Parameters used when defining a mitigation action that enable Amazon Web Services IoT Core logging.

ErrorInfo

Error information.

ExplicitDeny

Information that explicitly denies authorization.

ExponentialRolloutRate

Allows you to create an exponential rate of rollout for a job.

Field

Describes the name and data type at a field.

FileLocation

The location of the OTA update.

FirehoseAction

Describes an action that writes data to an Amazon Kinesis Firehose stream.

FleetMetricNameAndArn

The name and ARN of a fleet metric.

GroupNameAndArn

The name and ARN of a group.

HttpAction

Send data to an HTTPS endpoint.

HttpActionHeader

The HTTP action header.

HttpAuthorization

The authorization method used to send messages.

HttpContext

Specifies the HTTP context to use for the test authorizer request.

HttpUrlDestinationConfiguration

HTTP URL destination configuration used by the topic rule's HTTP action.

HttpUrlDestinationProperties

HTTP URL destination properties.

HttpUrlDestinationSummary

Information about an HTTP URL destination.

ImplicitDeny

Information that implicitly denies authorization. When policy doesn't explicitly deny

IotAnalyticsAction

Sends message data to an IoT Analytics channel.

IotEventsAction

Sends an input to an IoT Events detector.

IotSiteWiseAction

Describes an action to send data from an MQTT message that triggered the rule to IoT

Job

The Job object contains details about a job.

JobExecution

The job execution object represents the execution of a job on a particular device.

JobExecutionStatusDetails

Details of the job execution status.

JobExecutionSummary

The job execution summary.

JobExecutionSummaryForJob

Contains a summary of information about job executions for a specific job.

JobExecutionSummaryForThing

The job execution summary for a thing.

JobExecutionsRolloutConfig

Allows you to create a staged rollout of a job.

JobProcessDetails

The job process details.

JobSummary

The job summary.

JobTemplateSummary

An object that contains information about the job template.

KafkaAction

Send messages to an Amazon Managed Streaming for Apache Kafka (Amazon MSK) or self-managed Apache Kafka cluster.

KeyPair

Describes a key pair.

KinesisAction

Describes an action to write data to an Amazon Kinesis stream.

LambdaAction

Describes an action to invoke a Lambda function.

LogTarget

A log target.

LogTargetConfiguration

The target configuration.

LoggingOptionsPayload

Describes the logging options payload.

MachineLearningDetectionConfig

MetricDimension

The dimension of a metric.

MetricToRetain

The metric you want to retain. Dimensions are optional.

MetricValue

The value to be compared with the metric.

MitigationAction

Describes which changes should be applied as part of a mitigation action.

MitigationActionIdentifier

Information that identifies a mitigation action. This information is returned by ListMitigationActions.

MitigationActionParams

The set of parameters for this mitigation action. You can specify only one type of parameter (in other words, you can apply only one action for each defined mitigation action).

MqttContext

Specifies the MQTT context to use for the test authorizer request

NonCompliantResource

Information about the resource that was noncompliant with the audit check.

OpenSearchAction

Describes an action that writes data to an Amazon OpenSearch Service

OtaUpdateFile

Describes a file to be associated with an OTA update.

OtaUpdateInfo

Information about an OTA update.

OtaUpdateSummary

An OTA update summary.

OutgoingCertificate

A certificate that has been transferred but not yet accepted.

PercentPair

Describes the percentile and percentile value.

Policy

Describes an IoT policy.

PolicyVersion

Describes a policy version.

PolicyVersionIdentifier

Information about the version of the policy associated with the resource.

PresignedUrlConfig

Configuration for pre-signed S3 URLs.

ProvisioningHook

Structure that contains payloadVersion and

ProvisioningTemplateSummary

A summary of information about a fleet provisioning template.

ProvisioningTemplateVersionSummary

A summary of information about a fleet provision template version.

PublishFindingToSnsParams

Parameters to define a mitigation action that publishes findings to Amazon SNS. You can implement your own custom actions in response to the Amazon SNS messages.

PutAssetPropertyValueEntry

An asset property value entry containing the following information.

PutItemInput

The input for the DynamoActionVS action that specifies the DynamoDB table to which

RateIncreaseCriteria

Allows you to define a criteria to initiate the increase in rate of rollout for a job.

RegistrationConfig

The registration configuration.

RelatedResource

Information about a related resource.

ReplaceDefaultPolicyVersionParams

Parameters to define a mitigation action that adds a blank policy to restrict permissions.

RepublishAction

Describes an action to republish to another topic.

ResourceIdentifier

Information that identifies the noncompliant resource.

RoleAliasDescription

Role alias description.

S3Action

Describes an action to write data to an Amazon S3 bucket.

S3Destination

Describes the location of updated firmware in S3.

S3Location

The S3 location.

SalesforceAction

Describes an action to write a message to a Salesforce IoT Cloud Input

ScheduledAuditMetadata

Information about the scheduled audit.

SecurityProfileIdentifier

Identifying information for a Device Defender security profile.

SecurityProfileTarget

A target to which an alert is sent when a security profile behavior is

SecurityProfileTargetMapping

Information about a security profile and the target associated with it.

ServerCertificateSummary

An object that contains information about a server certificate.

SigV4Authorization

For more information, see Signature Version 4 signing process.

SigningProfileParameter

Describes the code-signing profile.

SnsAction

Describes an action to publish to an Amazon SNS topic.

SqsAction

Describes an action to publish data to an Amazon SQS queue.

StartSigningJobParameter

Information required to start a signing job.

StatisticalThreshold

A statistical ranking (percentile) that

Statistics

A map of key-value pairs for all supported statistics. Currently, only count is

StepFunctionsAction

Starts execution of a Step Functions state machine.

Stream

Describes a group of files that can be streamed.

StreamFile

Represents a file to stream.

StreamInfo

Information about a stream.

StreamSummary

A summary of a stream.

Tag

A set of key/value pairs that are used to manage the resource.

TaskStatistics

Statistics for the checks performed during the audit.

TaskStatisticsForAuditCheck

Provides summary counts of how many tasks for findings are in a particular state. This information is included in the response from DescribeAuditMitigationActionsTask.

TermsAggregation

Performs an aggregation that will return a list of buckets. The list of buckets is a ranked list of the number of occurrences of an aggregation field value.

ThingAttribute

The properties of the thing, including thing name, thing type name, and a list of thing

ThingConnectivity

The connectivity status of the thing.

ThingDocument

The thing search index document.

ThingGroupDocument

The thing group search index document.

ThingGroupIndexingConfiguration

Thing group indexing configuration.

ThingGroupMetadata

Thing group metadata.

ThingGroupProperties

Thing group properties.

ThingIndexingConfiguration

The thing indexing configuration. For more information, see Managing

ThingTypeDefinition

The definition of the thing type, including thing type name and description.

ThingTypeMetadata

The ThingTypeMetadata contains additional information about the thing type including: creation date and

ThingTypeProperties

The ThingTypeProperties contains information about the thing type including: a thing type description,

TimeoutConfig

Specifies the amount of time each device has to finish its execution of the job. A timer

TimestreamAction

The Timestream rule action writes attributes (measures) from an MQTT message

TimestreamDimension

Metadata attributes of the time series that are written in each measure record.

TimestreamTimestamp

Describes how to interpret an application-defined timestamp value from an MQTT message

TlsContext

Specifies the TLS context to use for the test authorizer request.

TopicRule

Describes a rule.

TopicRuleDestination

A topic rule destination.

TopicRuleDestinationConfiguration

Configuration of the topic rule destination.

TopicRuleDestinationSummary

Information about the topic rule destination.

TopicRuleListItem

Describes a rule.

TopicRulePayload

Describes a rule.

TransferData

Data used to transfer a certificate to an Amazon Web Services account.

UpdateCaCertificateParams

Parameters to define a mitigation action that changes the state of the CA certificate to inactive.

UpdateDeviceCertificateParams

Parameters to define a mitigation action that changes the state of the device certificate to inactive.

ValidationError

Information about an error found in a behavior specification.

ViolationEvent

Information about a Device Defender security profile behavior violation.

ViolationEventAdditionalInfo

ViolationEventOccurrenceRange

VpcDestinationConfiguration

The configuration information for a virtual private cloud (VPC) destination.

VpcDestinationProperties

The properties of a virtual private cloud (VPC) destination.

VpcDestinationSummary

The summary of a virtual private cloud (VPC) destination.

Enums

AbortAction

ActionType

AggregationTypeName

AlertTargetType

The type of alert target: one of "SNS".

AssetPropertyVariant

AuditCheckRunStatus

AuditFindingSeverity

AuditFrequency

AuditMitigationActionsExecutionStatus

AuditMitigationActionsTaskStatus

AuditNotificationType

AuditTaskStatus

AuditTaskType

AuthDecision

AuthorizerStatus

AutoRegistrationStatus

AwsJobAbortCriteriaAbortAction

AwsJobAbortCriteriaFailureType

BehaviorCriteriaType

CaCertificateStatus

CaCertificateUpdateAction

CannedAccessControlList

CertificateMode

CertificateStatus

ComparisonOperator

ConfidenceLevel

CustomMetricType

DayOfWeek

DetectMitigationActionExecutionStatus

DetectMitigationActionsTaskStatus

DeviceCertificateUpdateAction

DimensionType

DimensionValueOperator

DomainConfigurationStatus

DomainType

DynamicGroupStatus

DynamoKeyType

EventType

FieldType

FleetMetricUnit

IndexStatus

JobExecutionFailureType

JobExecutionStatus

JobStatus

LogLevel

LogTargetType

MessageFormat

MitigationActionType

ModelStatus

OtaUpdateStatus

PolicyTemplateName

Protocol

ReportType

ResourceType

ServerCertificateStatus

ServiceType

Status

TargetSelection

ThingConnectivityIndexingMode

ThingGroupIndexingMode

ThingIndexingMode

TopicRuleDestinationStatus

ViolationEventType