Struct aws_sdk_iot::model::action::Builder

#[non_exhaustive]
pub struct Builder { /* fields omitted */ }

A builder for Action

Implementations

impl Builder

pub fn dynamo_db(self, input: DynamoDbAction) -> Self

Write to a DynamoDB table.

pub fn set_dynamo_db(self, input: Option<DynamoDbAction>) -> Self

pub fn dynamo_d_bv2(self, input: DynamoDBv2Action) -> Self

Write to a DynamoDB table. This is a new version of the DynamoDB action. It allows you to write each attribute in an MQTT message payload into a separate DynamoDB column.

pub fn set_dynamo_d_bv2(self, input: Option<DynamoDBv2Action>) -> Self

pub fn lambda(self, input: LambdaAction) -> Self

Invoke a Lambda function.

pub fn set_lambda(self, input: Option<LambdaAction>) -> Self

pub fn sns(self, input: SnsAction) -> Self

Publish to an Amazon SNS topic.

pub fn set_sns(self, input: Option<SnsAction>) -> Self

pub fn sqs(self, input: SqsAction) -> Self

Publish to an Amazon SQS queue.

pub fn set_sqs(self, input: Option<SqsAction>) -> Self

pub fn kinesis(self, input: KinesisAction) -> Self

Write data to an Amazon Kinesis stream.

pub fn set_kinesis(self, input: Option<KinesisAction>) -> Self

pub fn republish(self, input: RepublishAction) -> Self

Publish to another MQTT topic.

pub fn set_republish(self, input: Option<RepublishAction>) -> Self

pub fn s3(self, input: S3Action) -> Self

Write to an Amazon S3 bucket.

pub fn set_s3(self, input: Option<S3Action>) -> Self

pub fn firehose(self, input: FirehoseAction) -> Self

Write to an Amazon Kinesis Firehose stream.

pub fn set_firehose(self, input: Option<FirehoseAction>) -> Self

pub fn cloudwatch_metric(self, input: CloudwatchMetricAction) -> Self

Capture a CloudWatch metric.

pub fn set_cloudwatch_metric(
self,
input: Option<CloudwatchMetricAction>
) -> Self

pub fn cloudwatch_alarm(self, input: CloudwatchAlarmAction) -> Self

Change the state of a CloudWatch alarm.

pub fn set_cloudwatch_alarm(self, input: Option<CloudwatchAlarmAction>) -> Self

pub fn cloudwatch_logs(self, input: CloudwatchLogsAction) -> Self

Send data to CloudWatch Logs.

pub fn set_cloudwatch_logs(self, input: Option<CloudwatchLogsAction>) -> Self

pub fn elasticsearch(self, input: ElasticsearchAction) -> Self

Write data to an Amazon Elasticsearch Service domain.

This action is deprecated. Use the OpenSearch action instead.

pub fn set_elasticsearch(self, input: Option<ElasticsearchAction>) -> Self

pub fn salesforce(self, input: SalesforceAction) -> Self

Send a message to a Salesforce IoT Cloud Input Stream.

pub fn set_salesforce(self, input: Option<SalesforceAction>) -> Self

pub fn iot_analytics(self, input: IotAnalyticsAction) -> Self

Sends message data to an IoT Analytics channel.

pub fn set_iot_analytics(self, input: Option<IotAnalyticsAction>) -> Self

pub fn iot_events(self, input: IotEventsAction) -> Self

Sends an input to an IoT Events detector.

pub fn set_iot_events(self, input: Option<IotEventsAction>) -> Self

pub fn iot_site_wise(self, input: IotSiteWiseAction) -> Self

Sends data from the MQTT message that triggered the rule to IoT SiteWise asset properties.

pub fn set_iot_site_wise(self, input: Option<IotSiteWiseAction>) -> Self

pub fn step_functions(self, input: StepFunctionsAction) -> Self

Starts execution of a Step Functions state machine.

pub fn set_step_functions(self, input: Option<StepFunctionsAction>) -> Self

pub fn timestream(self, input: TimestreamAction) -> Self

The Timestream rule action writes attributes (measures) from an MQTT message into an Amazon Timestream table. For more information, see the Timestream topic rule action documentation.

pub fn set_timestream(self, input: Option<TimestreamAction>) -> Self

pub fn http(self, input: HttpAction) -> Self

Send data to an HTTPS endpoint.

pub fn set_http(self, input: Option<HttpAction>) -> Self

pub fn kafka(self, input: KafkaAction) -> Self

Send messages to an Amazon Managed Streaming for Apache Kafka (Amazon MSK) or self-managed Apache Kafka cluster.

pub fn set_kafka(self, input: Option<KafkaAction>) -> Self

pub fn open_search(self, input: OpenSearchAction) -> Self

Write data to an Amazon OpenSearch Service domain.

pub fn set_open_search(self, input: Option<OpenSearchAction>) -> Self

pub fn build(self) -> Action

Consumes the builder and constructs a Action