Module aws_sdk_iot::operation

Structs

AcceptCertificateTransfer

Accepts a pending certificate transfer. The default state of the certificate is

AddThingToBillingGroup

Adds a thing to a billing group.

AddThingToThingGroup

Adds a thing to a thing group.

AssociateTargetsWithJob

Associates a group with a continuous job. The following criteria must be met:

AttachPolicy

Attaches the specified policy to the specified principal (certificate or other

AttachPrincipalPolicy

Attaches the specified policy to the specified principal (certificate or other

AttachSecurityProfile

Associates a Device Defender security profile with a thing group or this account. Each

AttachThingPrincipal

Attaches the specified principal to the specified thing. A principal can be X.509

CancelAuditMitigationActionsTask

Cancels a mitigation action task that is in progress. If the task

CancelAuditTask

Cancels an audit that is in progress. The audit can be either scheduled or on demand. If the audit isn't in progress, an "InvalidRequestException" occurs.

CancelCertificateTransfer

Cancels a pending transfer for the specified certificate.

CancelDetectMitigationActionsTask

CancelJob

Cancels a job.

CancelJobExecution

Cancels the execution of a job for a given thing.

ClearDefaultAuthorizer

Clears the default authorizer.

ConfirmTopicRuleDestination

Confirms a topic rule destination. When you create a rule requiring a destination, IoT

CreateAuditSuppression

CreateAuthorizer

Creates an authorizer.

CreateBillingGroup

Creates a billing group.

CreateCertificateFromCsr

Creates an X.509 certificate using the specified certificate signing

CreateCustomMetric

Use this API to define a

CreateDimension

Create a dimension that you can use to limit the scope of a metric used in a security profile for IoT Device Defender.

CreateDomainConfiguration

Creates a domain configuration.

CreateDynamicThingGroup

Creates a dynamic thing group.

CreateFleetMetric

Creates a fleet metric.

CreateJob

Creates a job.

CreateJobTemplate

Creates a job template.

CreateKeysAndCertificate

Creates a 2048-bit RSA key pair and issues an X.509 certificate using the issued

CreateMitigationAction

Defines an action that can be applied to audit findings by using StartAuditMitigationActionsTask. Only certain types of mitigation actions can be applied to specific check names.

CreateOTAUpdate

Creates an IoT OTA update on a target group of things or groups.

CreatePolicy

Creates an IoT policy.

CreatePolicyVersion

Creates a new version of the specified IoT policy. To update a policy, create a

CreateProvisioningClaim

Creates a provisioning claim.

CreateProvisioningTemplate

Creates a fleet provisioning template.

CreateProvisioningTemplateVersion

Creates a new version of a fleet provisioning template.

CreateRoleAlias

Creates a role alias.

CreateScheduledAudit

Creates a scheduled audit that is run at a specified

CreateSecurityProfile

Creates a Device Defender security profile.

CreateStream

Creates a stream for delivering one or more large files in chunks over MQTT. A stream transports data

CreateThing

Creates a thing record in the registry. If this call is made multiple times using

CreateThingGroup

Create a thing group.

CreateThingType

Creates a new thing type.

CreateTopicRule

Creates a rule. Creating rules is an administrator-level action. Any user who has

CreateTopicRuleDestination

Creates a topic rule destination. The destination must be confirmed prior to use.

DeleteAccountAuditConfiguration

Restores the default settings for Device Defender audits for this account. Any

DeleteAuditSuppression

DeleteAuthorizer

Deletes an authorizer.

DeleteBillingGroup

Deletes the billing group.

DeleteCACertificate

Deletes a registered CA certificate.

DeleteCertificate

Deletes the specified certificate.

DeleteCustomMetric

DeleteDimension

Removes the specified dimension from your Amazon Web Services accounts.

DeleteDomainConfiguration

Deletes the specified domain configuration.

DeleteDynamicThingGroup

Deletes a dynamic thing group.

DeleteFleetMetric

Deletes the specified fleet metric.

DeleteJob

Deletes a job and its related job executions.

DeleteJobExecution

Deletes a job execution.

DeleteJobTemplate

Deletes the specified job template.

DeleteMitigationAction

Deletes a defined mitigation action from your Amazon Web Services accounts.

DeleteOTAUpdate

Delete an OTA update.

DeletePolicy

Deletes the specified policy.

DeletePolicyVersion

Deletes the specified version of the specified policy. You cannot delete the default

DeleteProvisioningTemplate

Deletes a fleet provisioning template.

DeleteProvisioningTemplateVersion

Deletes a fleet provisioning template version.

DeleteRegistrationCode

Deletes a CA certificate registration code.

DeleteRoleAlias

Deletes a role alias

DeleteScheduledAudit

Deletes a scheduled audit.

DeleteSecurityProfile

Deletes a Device Defender security profile.

DeleteStream

Deletes a stream.

DeleteThing

Deletes the specified thing. Returns successfully with no error if the deletion is

DeleteThingGroup

Deletes a thing group.

DeleteThingType

Deletes the specified thing type. You cannot delete a thing type if it has things

DeleteTopicRule

Deletes the rule.

DeleteTopicRuleDestination

Deletes a topic rule destination.

DeleteV2LoggingLevel

Deletes a logging level.

DeprecateThingType

Deprecates a thing type. You can not associate new things with deprecated thing

DescribeAccountAuditConfiguration

Gets information about the Device Defender audit settings for this account.

DescribeAuditFinding

Gets information about a single audit finding. Properties include the reason for

DescribeAuditMitigationActionsTask

Gets information about an audit mitigation task that is used to apply mitigation actions to a set of audit findings. Properties include the actions being applied, the audit checks to which they're being applied, the task status, and aggregated task statistics.

DescribeAuditSuppression

DescribeAuditTask

Gets information about a Device Defender audit.

DescribeAuthorizer

Describes an authorizer.

DescribeBillingGroup

Returns information about a billing group.

DescribeCACertificate

Describes a registered CA certificate.

DescribeCertificate

Gets information about the specified certificate.

DescribeCustomMetric

DescribeDefaultAuthorizer

Describes the default authorizer.

DescribeDetectMitigationActionsTask

DescribeDimension

Provides details about a dimension that is defined in your Amazon Web Services accounts.

DescribeDomainConfiguration

Gets summary information about a domain configuration.

DescribeEndpoint

Returns a unique endpoint specific to the Amazon Web Services account making the call.

DescribeEventConfigurations

Describes event configurations.

DescribeFleetMetric

Gets information about the specified fleet metric.

DescribeIndex

Describes a search index.

DescribeJob

Describes a job.

DescribeJobExecution

Describes a job execution.

DescribeJobTemplate

Returns information about a job template.

DescribeMitigationAction

Gets information about a mitigation action.

DescribeProvisioningTemplate

Returns information about a fleet provisioning template.

DescribeProvisioningTemplateVersion

Returns information about a fleet provisioning template version.

DescribeRoleAlias

Describes a role alias.

DescribeScheduledAudit

Gets information about a scheduled audit.

DescribeSecurityProfile

Gets information about a Device Defender security profile.

DescribeStream

Gets information about a stream.

DescribeThing

Gets information about the specified thing.

DescribeThingGroup

Describe a thing group.

DescribeThingRegistrationTask

Describes a bulk thing provisioning task.

DescribeThingType

Gets information about the specified thing type.

DetachPolicy

Detaches a policy from the specified target.

DetachPrincipalPolicy

Removes the specified policy from the specified certificate.

DetachSecurityProfile

Disassociates a Device Defender security profile from a thing group or from this account.

DetachThingPrincipal

Detaches the specified principal from the specified thing. A principal can be X.509

DisableTopicRule

Disables the rule.

EnableTopicRule

Enables the rule.

GetBehaviorModelTrainingSummaries

GetBucketsAggregation

Aggregates on indexed data with search queries pertaining to particular fields.

GetCardinality

Returns the approximate count of unique values that match the query.

GetEffectivePolicies

Gets a list of the policies that have an effect on the authorization behavior of the

GetIndexingConfiguration

Gets the indexing configuration.

GetJobDocument

Gets a job document.

GetLoggingOptions

Gets the logging options.

GetOTAUpdate

Gets an OTA update.

GetPercentiles

Groups the aggregated values that match the query into percentile groupings. The default

GetPolicy

Gets information about the specified policy with the policy document of the default

GetPolicyVersion

Gets information about the specified policy version.

GetRegistrationCode

Gets a registration code used to register a CA certificate with IoT.

GetStatistics

Returns the count, average, sum, minimum, maximum, sum of squares, variance,

GetTopicRule

Gets information about the rule.

GetTopicRuleDestination

Gets information about a topic rule destination.

GetV2LoggingOptions

Gets the fine grained logging options.

ListActiveViolations

Lists the active violations for a given Device Defender security profile.

ListAttachedPolicies

Lists the policies attached to the specified thing group.

ListAuditFindings

Lists the findings (results) of a Device Defender audit or of the audits

ListAuditMitigationActionsExecutions

Gets the status of audit mitigation action tasks that were

ListAuditMitigationActionsTasks

Gets a list of audit mitigation action tasks that match the specified filters.

ListAuditSuppressions

ListAuditTasks

Lists the Device Defender audits that have been performed during a given

ListAuthorizers

Lists the authorizers registered in your account.

ListBillingGroups

Lists the billing groups you have created.

ListCACertificates

Lists the CA certificates registered for your Amazon Web Services account.

ListCertificates

Lists the certificates registered in your Amazon Web Services account.

ListCertificatesByCA

List the device certificates signed by the specified CA certificate.

ListCustomMetrics

ListDetectMitigationActionsExecutions

ListDetectMitigationActionsTasks

ListDimensions

List the set of dimensions that are defined for your Amazon Web Services accounts.

ListDomainConfigurations

Gets a list of domain configurations for the user. This list is sorted

ListFleetMetrics

Lists all your fleet metrics.

ListIndices

Lists the search indices.

ListJobExecutionsForJob

Lists the job executions for a job.

ListJobExecutionsForThing

Lists the job executions for the specified thing.

ListJobTemplates

Returns a list of job templates.

ListJobs

Lists jobs.

ListMitigationActions

Gets a list of all mitigation actions that match the specified filter criteria.

ListOTAUpdates

Lists OTA updates.

ListOutgoingCertificates

Lists certificates that are being transferred but not yet accepted.

ListPolicies

Lists your policies.

ListPolicyPrincipals

Lists the principals associated with the specified policy.

ListPolicyVersions

Lists the versions of the specified policy and identifies the default

ListPrincipalPolicies

Lists the policies attached to the specified principal. If you use an Cognito

ListPrincipalThings

Lists the things associated with the specified principal. A principal can be X.509

ListProvisioningTemplateVersions

A list of fleet provisioning template versions.

ListProvisioningTemplates

Lists the fleet provisioning templates in your Amazon Web Services account.

ListRoleAliases

Lists the role aliases registered in your account.

ListScheduledAudits

Lists all of your scheduled audits.

ListSecurityProfiles

Lists the Device Defender security profiles

ListSecurityProfilesForTarget

Lists the Device Defender security profiles attached to a target (thing group).

ListStreams

Lists all of the streams in your Amazon Web Services account.

ListTagsForResource

Lists the tags (metadata) you have assigned to the resource.

ListTargetsForPolicy

List targets for the specified policy.

ListTargetsForSecurityProfile

Lists the targets (thing groups) associated with a given Device Defender security profile.

ListThingGroups

List the thing groups in your account.

ListThingGroupsForThing

List the thing groups to which the specified thing belongs.

ListThingPrincipals

Lists the principals associated with the specified thing. A principal can be X.509

ListThingRegistrationTaskReports

Information about the thing registration tasks.

ListThingRegistrationTasks

List bulk thing provisioning tasks.

ListThingTypes

Lists the existing thing types.

ListThings

Lists your things. Use the attributeName and

ListThingsInBillingGroup

Lists the things you have added to the given billing group.

ListThingsInThingGroup

Lists the things in the specified group.

ListTopicRuleDestinations

Lists all the topic rule destinations in your Amazon Web Services account.

ListTopicRules

Lists the rules for the specific topic.

ListV2LoggingLevels

Lists logging levels.

ListViolationEvents

Lists the Device Defender security profile violations discovered during the given time period.

RegisterCACertificate

Registers a CA certificate with IoT. This CA certificate can then be used to sign

RegisterCertificate

Registers a device certificate with IoT. If you have more than one CA certificate

RegisterCertificateWithoutCA

Register a certificate that does not have a certificate authority (CA).

RegisterThing

Provisions a thing in the device registry. RegisterThing calls other IoT control

RejectCertificateTransfer

Rejects a pending certificate transfer. After IoT rejects a certificate transfer,

RemoveThingFromBillingGroup

Removes the given thing from the billing group.

RemoveThingFromThingGroup

Remove the specified thing from the specified group.

ReplaceTopicRule

Replaces the rule. You must specify all parameters for the new rule. Creating rules

SearchIndex

The query search index.

SetDefaultAuthorizer

Sets the default authorizer. This will be used if a websocket connection is made

SetDefaultPolicyVersion

Sets the specified version of the specified policy as the policy's default

SetLoggingOptions

Sets the logging options.

SetV2LoggingLevel

Sets the logging level.

SetV2LoggingOptions

Sets the logging options for the V2 logging service.

StartAuditMitigationActionsTask

Starts a task that applies a set of mitigation actions to the specified target.

StartDetectMitigationActionsTask

StartOnDemandAuditTask

Starts an on-demand Device Defender audit.

StartThingRegistrationTask

Creates a bulk thing provisioning task.

StopThingRegistrationTask

Cancels a bulk thing provisioning task.

TagResource

Adds to or modifies the tags of the given resource. Tags are metadata which can be

TestAuthorization

Tests if a specified principal is authorized to perform an IoT action on a

TestInvokeAuthorizer

Tests a custom authorization behavior by invoking a specified custom authorizer. Use

TransferCertificate

Transfers the specified certificate to the specified Amazon Web Services account.

UntagResource

Removes the given tags (metadata) from the resource.

UpdateAccountAuditConfiguration

Configures or reconfigures the Device Defender audit settings for this account.

UpdateAuditSuppression

UpdateAuthorizer

Updates an authorizer.

UpdateBillingGroup

Updates information about the billing group.

UpdateCACertificate

Updates a registered CA certificate.

UpdateCertificate

Updates the status of the specified certificate. This operation is

UpdateCustomMetric

Updates a

UpdateDimension

Updates the definition for a dimension. You

UpdateDomainConfiguration

Updates values stored in the domain configuration. Domain configurations for default

UpdateDynamicThingGroup

Updates a dynamic thing group.

UpdateEventConfigurations

Updates the event configurations.

UpdateFleetMetric

Updates the data for a fleet metric.

UpdateIndexingConfiguration

Updates the search configuration.

UpdateJob

Updates supported fields of the specified job.

UpdateMitigationAction

Updates the definition for the specified mitigation action.

UpdateProvisioningTemplate

Updates a fleet provisioning template.

UpdateRoleAlias

Updates a role alias.

UpdateScheduledAudit

Updates a scheduled audit, including which checks are performed and

UpdateSecurityProfile

Updates a Device Defender security profile.

UpdateStream

Updates an existing stream. The stream version will be incremented by one.

UpdateThing

Updates the data for a thing.

UpdateThingGroup

Update a thing group.

UpdateThingGroupsForThing

Updates the groups to which the thing belongs.

UpdateTopicRuleDestination

Updates a topic rule destination. You use this to change the status, endpoint URL, or

ValidateSecurityProfileBehaviors

Validates a Device Defender security profile behaviors specification.