Struct aws_sdk_iot::client::Client

pub struct Client<C = DynConnector, M = AwsMiddleware, R = Standard> { /* fields omitted */ }

An ergonomic service client for AWSIotService.

This client allows ergonomic access to a AWSIotService-shaped service. Each method corresponds to an endpoint defined in the service’s Smithy model, and the request and response shapes are auto-generated from that same model.

Using a Client

Once you have a client set up, you can access the service’s endpoints by calling the appropriate method on Client. Each such method returns a request builder for that endpoint, with methods for setting the various fields of the request. Once your request is complete, use the send method to send the request. send returns a future, which you then have to .await to get the service’s response.

Implementations

impl<C, M, R> Client<C, M, R>

pub fn with_config(client: Client<C, M, R>, conf: Config) -> Self

Creates a client with the given service configuration.

pub fn conf(&self) -> &Config

Returns the client’s configuration.

impl<C, M, R> Client<C, M, R> where
C: SmithyConnector,
M: SmithyMiddleware<C>,
R: NewRequestPolicy,

pub fn accept_certificate_transfer(&self) -> AcceptCertificateTransfer<C, M, R>

Constructs a fluent builder for the AcceptCertificateTransfer operation.

See AcceptCertificateTransfer for more information about the operation and its arguments.

pub fn add_thing_to_billing_group(&self) -> AddThingToBillingGroup<C, M, R>

Constructs a fluent builder for the AddThingToBillingGroup operation.

See AddThingToBillingGroup for more information about the operation and its arguments.

pub fn add_thing_to_thing_group(&self) -> AddThingToThingGroup<C, M, R>

Constructs a fluent builder for the AddThingToThingGroup operation.

See AddThingToThingGroup for more information about the operation and its arguments.

pub fn associate_targets_with_job(&self) -> AssociateTargetsWithJob<C, M, R>

Constructs a fluent builder for the AssociateTargetsWithJob operation.

See AssociateTargetsWithJob for more information about the operation and its arguments.

pub fn attach_policy(&self) -> AttachPolicy<C, M, R>

Constructs a fluent builder for the AttachPolicy operation.

See AttachPolicy for more information about the operation and its arguments.

pub fn attach_principal_policy(&self) -> AttachPrincipalPolicy<C, M, R>

Constructs a fluent builder for the AttachPrincipalPolicy operation.

See AttachPrincipalPolicy for more information about the operation and its arguments.

pub fn attach_security_profile(&self) -> AttachSecurityProfile<C, M, R>

Constructs a fluent builder for the AttachSecurityProfile operation.

See AttachSecurityProfile for more information about the operation and its arguments.

pub fn attach_thing_principal(&self) -> AttachThingPrincipal<C, M, R>

Constructs a fluent builder for the AttachThingPrincipal operation.

See AttachThingPrincipal for more information about the operation and its arguments.

pub fn cancel_audit_mitigation_actions_task(
&self
) -> CancelAuditMitigationActionsTask<C, M, R>

Constructs a fluent builder for the CancelAuditMitigationActionsTask operation.

See CancelAuditMitigationActionsTask for more information about the operation and its arguments.

pub fn cancel_audit_task(&self) -> CancelAuditTask<C, M, R>

Constructs a fluent builder for the CancelAuditTask operation.

See CancelAuditTask for more information about the operation and its arguments.

pub fn cancel_certificate_transfer(&self) -> CancelCertificateTransfer<C, M, R>

Constructs a fluent builder for the CancelCertificateTransfer operation.

See CancelCertificateTransfer for more information about the operation and its arguments.

pub fn cancel_detect_mitigation_actions_task(
&self
) -> CancelDetectMitigationActionsTask<C, M, R>

Constructs a fluent builder for the CancelDetectMitigationActionsTask operation.

See CancelDetectMitigationActionsTask for more information about the operation and its arguments.

pub fn cancel_job(&self) -> CancelJob<C, M, R>

Constructs a fluent builder for the CancelJob operation.

See CancelJob for more information about the operation and its arguments.

pub fn cancel_job_execution(&self) -> CancelJobExecution<C, M, R>

Constructs a fluent builder for the CancelJobExecution operation.

See CancelJobExecution for more information about the operation and its arguments.

pub fn clear_default_authorizer(&self) -> ClearDefaultAuthorizer<C, M, R>

Constructs a fluent builder for the ClearDefaultAuthorizer operation.

See ClearDefaultAuthorizer for more information about the operation and its arguments.

pub fn confirm_topic_rule_destination(
&self
) -> ConfirmTopicRuleDestination<C, M, R>

Constructs a fluent builder for the ConfirmTopicRuleDestination operation.

See ConfirmTopicRuleDestination for more information about the operation and its arguments.

pub fn create_audit_suppression(&self) -> CreateAuditSuppression<C, M, R>

Constructs a fluent builder for the CreateAuditSuppression operation.

See CreateAuditSuppression for more information about the operation and its arguments.

pub fn create_authorizer(&self) -> CreateAuthorizer<C, M, R>

Constructs a fluent builder for the CreateAuthorizer operation.

See CreateAuthorizer for more information about the operation and its arguments.

pub fn create_billing_group(&self) -> CreateBillingGroup<C, M, R>

Constructs a fluent builder for the CreateBillingGroup operation.

See CreateBillingGroup for more information about the operation and its arguments.

pub fn create_certificate_from_csr(&self) -> CreateCertificateFromCsr<C, M, R>

Constructs a fluent builder for the CreateCertificateFromCsr operation.

See CreateCertificateFromCsr for more information about the operation and its arguments.

pub fn create_custom_metric(&self) -> CreateCustomMetric<C, M, R>

Constructs a fluent builder for the CreateCustomMetric operation.

See CreateCustomMetric for more information about the operation and its arguments.

pub fn create_dimension(&self) -> CreateDimension<C, M, R>

Constructs a fluent builder for the CreateDimension operation.

See CreateDimension for more information about the operation and its arguments.

pub fn create_domain_configuration(&self) -> CreateDomainConfiguration<C, M, R>

Constructs a fluent builder for the CreateDomainConfiguration operation.

See CreateDomainConfiguration for more information about the operation and its arguments.

pub fn create_dynamic_thing_group(&self) -> CreateDynamicThingGroup<C, M, R>

Constructs a fluent builder for the CreateDynamicThingGroup operation.

See CreateDynamicThingGroup for more information about the operation and its arguments.

pub fn create_fleet_metric(&self) -> CreateFleetMetric<C, M, R>

Constructs a fluent builder for the CreateFleetMetric operation.

See CreateFleetMetric for more information about the operation and its arguments.

pub fn create_job(&self) -> CreateJob<C, M, R>

Constructs a fluent builder for the CreateJob operation.

See CreateJob for more information about the operation and its arguments.

pub fn create_job_template(&self) -> CreateJobTemplate<C, M, R>

Constructs a fluent builder for the CreateJobTemplate operation.

See CreateJobTemplate for more information about the operation and its arguments.

pub fn create_keys_and_certificate(&self) -> CreateKeysAndCertificate<C, M, R>

Constructs a fluent builder for the CreateKeysAndCertificate operation.

See CreateKeysAndCertificate for more information about the operation and its arguments.

pub fn create_mitigation_action(&self) -> CreateMitigationAction<C, M, R>

Constructs a fluent builder for the CreateMitigationAction operation.

See CreateMitigationAction for more information about the operation and its arguments.

pub fn create_ota_update(&self) -> CreateOTAUpdate<C, M, R>

Constructs a fluent builder for the CreateOTAUpdate operation.

See CreateOTAUpdate for more information about the operation and its arguments.

pub fn create_policy(&self) -> CreatePolicy<C, M, R>

Constructs a fluent builder for the CreatePolicy operation.

See CreatePolicy for more information about the operation and its arguments.

pub fn create_policy_version(&self) -> CreatePolicyVersion<C, M, R>

Constructs a fluent builder for the CreatePolicyVersion operation.

See CreatePolicyVersion for more information about the operation and its arguments.

pub fn create_provisioning_claim(&self) -> CreateProvisioningClaim<C, M, R>

Constructs a fluent builder for the CreateProvisioningClaim operation.

See CreateProvisioningClaim for more information about the operation and its arguments.

pub fn create_provisioning_template(
&self
) -> CreateProvisioningTemplate<C, M, R>

Constructs a fluent builder for the CreateProvisioningTemplate operation.

See CreateProvisioningTemplate for more information about the operation and its arguments.

pub fn create_provisioning_template_version(
&self
) -> CreateProvisioningTemplateVersion<C, M, R>

Constructs a fluent builder for the CreateProvisioningTemplateVersion operation.

See CreateProvisioningTemplateVersion for more information about the operation and its arguments.

pub fn create_role_alias(&self) -> CreateRoleAlias<C, M, R>

Constructs a fluent builder for the CreateRoleAlias operation.

See CreateRoleAlias for more information about the operation and its arguments.

pub fn create_scheduled_audit(&self) -> CreateScheduledAudit<C, M, R>

Constructs a fluent builder for the CreateScheduledAudit operation.

See CreateScheduledAudit for more information about the operation and its arguments.

pub fn create_security_profile(&self) -> CreateSecurityProfile<C, M, R>

Constructs a fluent builder for the CreateSecurityProfile operation.

See CreateSecurityProfile for more information about the operation and its arguments.

pub fn create_stream(&self) -> CreateStream<C, M, R>

Constructs a fluent builder for the CreateStream operation.

See CreateStream for more information about the operation and its arguments.

pub fn create_thing(&self) -> CreateThing<C, M, R>

Constructs a fluent builder for the CreateThing operation.

See CreateThing for more information about the operation and its arguments.

pub fn create_thing_group(&self) -> CreateThingGroup<C, M, R>

Constructs a fluent builder for the CreateThingGroup operation.

See CreateThingGroup for more information about the operation and its arguments.

pub fn create_thing_type(&self) -> CreateThingType<C, M, R>

Constructs a fluent builder for the CreateThingType operation.

See CreateThingType for more information about the operation and its arguments.

pub fn create_topic_rule(&self) -> CreateTopicRule<C, M, R>

Constructs a fluent builder for the CreateTopicRule operation.

See CreateTopicRule for more information about the operation and its arguments.

pub fn create_topic_rule_destination(
&self
) -> CreateTopicRuleDestination<C, M, R>

Constructs a fluent builder for the CreateTopicRuleDestination operation.

See CreateTopicRuleDestination for more information about the operation and its arguments.

pub fn delete_account_audit_configuration(
&self
) -> DeleteAccountAuditConfiguration<C, M, R>

Constructs a fluent builder for the DeleteAccountAuditConfiguration operation.

See DeleteAccountAuditConfiguration for more information about the operation and its arguments.

pub fn delete_audit_suppression(&self) -> DeleteAuditSuppression<C, M, R>

Constructs a fluent builder for the DeleteAuditSuppression operation.

See DeleteAuditSuppression for more information about the operation and its arguments.

pub fn delete_authorizer(&self) -> DeleteAuthorizer<C, M, R>

Constructs a fluent builder for the DeleteAuthorizer operation.

See DeleteAuthorizer for more information about the operation and its arguments.

pub fn delete_billing_group(&self) -> DeleteBillingGroup<C, M, R>

Constructs a fluent builder for the DeleteBillingGroup operation.

See DeleteBillingGroup for more information about the operation and its arguments.

pub fn delete_ca_certificate(&self) -> DeleteCACertificate<C, M, R>

Constructs a fluent builder for the DeleteCACertificate operation.

See DeleteCACertificate for more information about the operation and its arguments.

pub fn delete_certificate(&self) -> DeleteCertificate<C, M, R>

Constructs a fluent builder for the DeleteCertificate operation.

See DeleteCertificate for more information about the operation and its arguments.

pub fn delete_custom_metric(&self) -> DeleteCustomMetric<C, M, R>

Constructs a fluent builder for the DeleteCustomMetric operation.

See DeleteCustomMetric for more information about the operation and its arguments.

pub fn delete_dimension(&self) -> DeleteDimension<C, M, R>

Constructs a fluent builder for the DeleteDimension operation.

See DeleteDimension for more information about the operation and its arguments.

pub fn delete_domain_configuration(&self) -> DeleteDomainConfiguration<C, M, R>

Constructs a fluent builder for the DeleteDomainConfiguration operation.

See DeleteDomainConfiguration for more information about the operation and its arguments.

pub fn delete_dynamic_thing_group(&self) -> DeleteDynamicThingGroup<C, M, R>

Constructs a fluent builder for the DeleteDynamicThingGroup operation.

See DeleteDynamicThingGroup for more information about the operation and its arguments.

pub fn delete_fleet_metric(&self) -> DeleteFleetMetric<C, M, R>

Constructs a fluent builder for the DeleteFleetMetric operation.

See DeleteFleetMetric for more information about the operation and its arguments.

pub fn delete_job(&self) -> DeleteJob<C, M, R>

Constructs a fluent builder for the DeleteJob operation.

See DeleteJob for more information about the operation and its arguments.

pub fn delete_job_execution(&self) -> DeleteJobExecution<C, M, R>

Constructs a fluent builder for the DeleteJobExecution operation.

See DeleteJobExecution for more information about the operation and its arguments.

pub fn delete_job_template(&self) -> DeleteJobTemplate<C, M, R>

Constructs a fluent builder for the DeleteJobTemplate operation.

See DeleteJobTemplate for more information about the operation and its arguments.

pub fn delete_mitigation_action(&self) -> DeleteMitigationAction<C, M, R>

Constructs a fluent builder for the DeleteMitigationAction operation.

See DeleteMitigationAction for more information about the operation and its arguments.

pub fn delete_ota_update(&self) -> DeleteOTAUpdate<C, M, R>

Constructs a fluent builder for the DeleteOTAUpdate operation.

See DeleteOTAUpdate for more information about the operation and its arguments.

pub fn delete_policy(&self) -> DeletePolicy<C, M, R>

Constructs a fluent builder for the DeletePolicy operation.

See DeletePolicy for more information about the operation and its arguments.

pub fn delete_policy_version(&self) -> DeletePolicyVersion<C, M, R>

Constructs a fluent builder for the DeletePolicyVersion operation.

See DeletePolicyVersion for more information about the operation and its arguments.

pub fn delete_provisioning_template(
&self
) -> DeleteProvisioningTemplate<C, M, R>

Constructs a fluent builder for the DeleteProvisioningTemplate operation.

See DeleteProvisioningTemplate for more information about the operation and its arguments.

pub fn delete_provisioning_template_version(
&self
) -> DeleteProvisioningTemplateVersion<C, M, R>

Constructs a fluent builder for the DeleteProvisioningTemplateVersion operation.

See DeleteProvisioningTemplateVersion for more information about the operation and its arguments.

pub fn delete_registration_code(&self) -> DeleteRegistrationCode<C, M, R>

Constructs a fluent builder for the DeleteRegistrationCode operation.

See DeleteRegistrationCode for more information about the operation and its arguments.

pub fn delete_role_alias(&self) -> DeleteRoleAlias<C, M, R>

Constructs a fluent builder for the DeleteRoleAlias operation.

See DeleteRoleAlias for more information about the operation and its arguments.

pub fn delete_scheduled_audit(&self) -> DeleteScheduledAudit<C, M, R>

Constructs a fluent builder for the DeleteScheduledAudit operation.

See DeleteScheduledAudit for more information about the operation and its arguments.

pub fn delete_security_profile(&self) -> DeleteSecurityProfile<C, M, R>

Constructs a fluent builder for the DeleteSecurityProfile operation.

See DeleteSecurityProfile for more information about the operation and its arguments.

pub fn delete_stream(&self) -> DeleteStream<C, M, R>

Constructs a fluent builder for the DeleteStream operation.

See DeleteStream for more information about the operation and its arguments.

pub fn delete_thing(&self) -> DeleteThing<C, M, R>

Constructs a fluent builder for the DeleteThing operation.

See DeleteThing for more information about the operation and its arguments.

pub fn delete_thing_group(&self) -> DeleteThingGroup<C, M, R>

Constructs a fluent builder for the DeleteThingGroup operation.

See DeleteThingGroup for more information about the operation and its arguments.

pub fn delete_thing_type(&self) -> DeleteThingType<C, M, R>

Constructs a fluent builder for the DeleteThingType operation.

See DeleteThingType for more information about the operation and its arguments.

pub fn delete_topic_rule(&self) -> DeleteTopicRule<C, M, R>

Constructs a fluent builder for the DeleteTopicRule operation.

See DeleteTopicRule for more information about the operation and its arguments.

pub fn delete_topic_rule_destination(
&self
) -> DeleteTopicRuleDestination<C, M, R>

Constructs a fluent builder for the DeleteTopicRuleDestination operation.

See DeleteTopicRuleDestination for more information about the operation and its arguments.

pub fn delete_v2_logging_level(&self) -> DeleteV2LoggingLevel<C, M, R>

Constructs a fluent builder for the DeleteV2LoggingLevel operation.

See DeleteV2LoggingLevel for more information about the operation and its arguments.

pub fn deprecate_thing_type(&self) -> DeprecateThingType<C, M, R>

Constructs a fluent builder for the DeprecateThingType operation.

See DeprecateThingType for more information about the operation and its arguments.

pub fn describe_account_audit_configuration(
&self
) -> DescribeAccountAuditConfiguration<C, M, R>

Constructs a fluent builder for the DescribeAccountAuditConfiguration operation.

See DescribeAccountAuditConfiguration for more information about the operation and its arguments.

pub fn describe_audit_finding(&self) -> DescribeAuditFinding<C, M, R>

Constructs a fluent builder for the DescribeAuditFinding operation.

See DescribeAuditFinding for more information about the operation and its arguments.

pub fn describe_audit_mitigation_actions_task(
&self
) -> DescribeAuditMitigationActionsTask<C, M, R>

Constructs a fluent builder for the DescribeAuditMitigationActionsTask operation.

See DescribeAuditMitigationActionsTask for more information about the operation and its arguments.

pub fn describe_audit_suppression(&self) -> DescribeAuditSuppression<C, M, R>

Constructs a fluent builder for the DescribeAuditSuppression operation.

See DescribeAuditSuppression for more information about the operation and its arguments.

pub fn describe_audit_task(&self) -> DescribeAuditTask<C, M, R>

Constructs a fluent builder for the DescribeAuditTask operation.

See DescribeAuditTask for more information about the operation and its arguments.

pub fn describe_authorizer(&self) -> DescribeAuthorizer<C, M, R>

Constructs a fluent builder for the DescribeAuthorizer operation.

See DescribeAuthorizer for more information about the operation and its arguments.

pub fn describe_billing_group(&self) -> DescribeBillingGroup<C, M, R>

Constructs a fluent builder for the DescribeBillingGroup operation.

See DescribeBillingGroup for more information about the operation and its arguments.

pub fn describe_ca_certificate(&self) -> DescribeCACertificate<C, M, R>

Constructs a fluent builder for the DescribeCACertificate operation.

See DescribeCACertificate for more information about the operation and its arguments.

pub fn describe_certificate(&self) -> DescribeCertificate<C, M, R>

Constructs a fluent builder for the DescribeCertificate operation.

See DescribeCertificate for more information about the operation and its arguments.

pub fn describe_custom_metric(&self) -> DescribeCustomMetric<C, M, R>

Constructs a fluent builder for the DescribeCustomMetric operation.

See DescribeCustomMetric for more information about the operation and its arguments.

pub fn describe_default_authorizer(&self) -> DescribeDefaultAuthorizer<C, M, R>

Constructs a fluent builder for the DescribeDefaultAuthorizer operation.

See DescribeDefaultAuthorizer for more information about the operation and its arguments.

pub fn describe_detect_mitigation_actions_task(
&self
) -> DescribeDetectMitigationActionsTask<C, M, R>

Constructs a fluent builder for the DescribeDetectMitigationActionsTask operation.

See DescribeDetectMitigationActionsTask for more information about the operation and its arguments.

pub fn describe_dimension(&self) -> DescribeDimension<C, M, R>

Constructs a fluent builder for the DescribeDimension operation.

See DescribeDimension for more information about the operation and its arguments.

pub fn describe_domain_configuration(
&self
) -> DescribeDomainConfiguration<C, M, R>

Constructs a fluent builder for the DescribeDomainConfiguration operation.

See DescribeDomainConfiguration for more information about the operation and its arguments.

pub fn describe_endpoint(&self) -> DescribeEndpoint<C, M, R>

Constructs a fluent builder for the DescribeEndpoint operation.

See DescribeEndpoint for more information about the operation and its arguments.

pub fn describe_event_configurations(
&self
) -> DescribeEventConfigurations<C, M, R>

Constructs a fluent builder for the DescribeEventConfigurations operation.

See DescribeEventConfigurations for more information about the operation and its arguments.

pub fn describe_fleet_metric(&self) -> DescribeFleetMetric<C, M, R>

Constructs a fluent builder for the DescribeFleetMetric operation.

See DescribeFleetMetric for more information about the operation and its arguments.

pub fn describe_index(&self) -> DescribeIndex<C, M, R>

Constructs a fluent builder for the DescribeIndex operation.

See DescribeIndex for more information about the operation and its arguments.

pub fn describe_job(&self) -> DescribeJob<C, M, R>

Constructs a fluent builder for the DescribeJob operation.

See DescribeJob for more information about the operation and its arguments.

pub fn describe_job_execution(&self) -> DescribeJobExecution<C, M, R>

Constructs a fluent builder for the DescribeJobExecution operation.

See DescribeJobExecution for more information about the operation and its arguments.

pub fn describe_job_template(&self) -> DescribeJobTemplate<C, M, R>

Constructs a fluent builder for the DescribeJobTemplate operation.

See DescribeJobTemplate for more information about the operation and its arguments.

pub fn describe_mitigation_action(&self) -> DescribeMitigationAction<C, M, R>

Constructs a fluent builder for the DescribeMitigationAction operation.

See DescribeMitigationAction for more information about the operation and its arguments.

pub fn describe_provisioning_template(
&self
) -> DescribeProvisioningTemplate<C, M, R>

Constructs a fluent builder for the DescribeProvisioningTemplate operation.

See DescribeProvisioningTemplate for more information about the operation and its arguments.

pub fn describe_provisioning_template_version(
&self
) -> DescribeProvisioningTemplateVersion<C, M, R>

Constructs a fluent builder for the DescribeProvisioningTemplateVersion operation.

See DescribeProvisioningTemplateVersion for more information about the operation and its arguments.

pub fn describe_role_alias(&self) -> DescribeRoleAlias<C, M, R>

Constructs a fluent builder for the DescribeRoleAlias operation.

See DescribeRoleAlias for more information about the operation and its arguments.

pub fn describe_scheduled_audit(&self) -> DescribeScheduledAudit<C, M, R>

Constructs a fluent builder for the DescribeScheduledAudit operation.

See DescribeScheduledAudit for more information about the operation and its arguments.

pub fn describe_security_profile(&self) -> DescribeSecurityProfile<C, M, R>

Constructs a fluent builder for the DescribeSecurityProfile operation.

See DescribeSecurityProfile for more information about the operation and its arguments.

pub fn describe_stream(&self) -> DescribeStream<C, M, R>

Constructs a fluent builder for the DescribeStream operation.

See DescribeStream for more information about the operation and its arguments.

pub fn describe_thing(&self) -> DescribeThing<C, M, R>

Constructs a fluent builder for the DescribeThing operation.

See DescribeThing for more information about the operation and its arguments.

pub fn describe_thing_group(&self) -> DescribeThingGroup<C, M, R>

Constructs a fluent builder for the DescribeThingGroup operation.

See DescribeThingGroup for more information about the operation and its arguments.

pub fn describe_thing_registration_task(
&self
) -> DescribeThingRegistrationTask<C, M, R>

Constructs a fluent builder for the DescribeThingRegistrationTask operation.

See DescribeThingRegistrationTask for more information about the operation and its arguments.

pub fn describe_thing_type(&self) -> DescribeThingType<C, M, R>

Constructs a fluent builder for the DescribeThingType operation.

See DescribeThingType for more information about the operation and its arguments.

pub fn detach_policy(&self) -> DetachPolicy<C, M, R>

Constructs a fluent builder for the DetachPolicy operation.

See DetachPolicy for more information about the operation and its arguments.

pub fn detach_principal_policy(&self) -> DetachPrincipalPolicy<C, M, R>

Constructs a fluent builder for the DetachPrincipalPolicy operation.

See DetachPrincipalPolicy for more information about the operation and its arguments.

pub fn detach_security_profile(&self) -> DetachSecurityProfile<C, M, R>

Constructs a fluent builder for the DetachSecurityProfile operation.

See DetachSecurityProfile for more information about the operation and its arguments.

pub fn detach_thing_principal(&self) -> DetachThingPrincipal<C, M, R>

Constructs a fluent builder for the DetachThingPrincipal operation.

See DetachThingPrincipal for more information about the operation and its arguments.

pub fn disable_topic_rule(&self) -> DisableTopicRule<C, M, R>

Constructs a fluent builder for the DisableTopicRule operation.

See DisableTopicRule for more information about the operation and its arguments.

pub fn enable_topic_rule(&self) -> EnableTopicRule<C, M, R>

Constructs a fluent builder for the EnableTopicRule operation.

See EnableTopicRule for more information about the operation and its arguments.

pub fn get_behavior_model_training_summaries(
&self
) -> GetBehaviorModelTrainingSummaries<C, M, R>

Constructs a fluent builder for the GetBehaviorModelTrainingSummaries operation.

See GetBehaviorModelTrainingSummaries for more information about the operation and its arguments.

pub fn get_buckets_aggregation(&self) -> GetBucketsAggregation<C, M, R>

Constructs a fluent builder for the GetBucketsAggregation operation.

See GetBucketsAggregation for more information about the operation and its arguments.

pub fn get_cardinality(&self) -> GetCardinality<C, M, R>

Constructs a fluent builder for the GetCardinality operation.

See GetCardinality for more information about the operation and its arguments.

pub fn get_effective_policies(&self) -> GetEffectivePolicies<C, M, R>

Constructs a fluent builder for the GetEffectivePolicies operation.

See GetEffectivePolicies for more information about the operation and its arguments.

pub fn get_indexing_configuration(&self) -> GetIndexingConfiguration<C, M, R>

Constructs a fluent builder for the GetIndexingConfiguration operation.

See GetIndexingConfiguration for more information about the operation and its arguments.

pub fn get_job_document(&self) -> GetJobDocument<C, M, R>

Constructs a fluent builder for the GetJobDocument operation.

See GetJobDocument for more information about the operation and its arguments.

pub fn get_logging_options(&self) -> GetLoggingOptions<C, M, R>

Constructs a fluent builder for the GetLoggingOptions operation.

See GetLoggingOptions for more information about the operation and its arguments.

pub fn get_ota_update(&self) -> GetOTAUpdate<C, M, R>

Constructs a fluent builder for the GetOTAUpdate operation.

See GetOTAUpdate for more information about the operation and its arguments.

pub fn get_percentiles(&self) -> GetPercentiles<C, M, R>

Constructs a fluent builder for the GetPercentiles operation.

See GetPercentiles for more information about the operation and its arguments.

pub fn get_policy(&self) -> GetPolicy<C, M, R>

Constructs a fluent builder for the GetPolicy operation.

See GetPolicy for more information about the operation and its arguments.

pub fn get_policy_version(&self) -> GetPolicyVersion<C, M, R>

Constructs a fluent builder for the GetPolicyVersion operation.

See GetPolicyVersion for more information about the operation and its arguments.

pub fn get_registration_code(&self) -> GetRegistrationCode<C, M, R>

Constructs a fluent builder for the GetRegistrationCode operation.

See GetRegistrationCode for more information about the operation and its arguments.

pub fn get_statistics(&self) -> GetStatistics<C, M, R>

Constructs a fluent builder for the GetStatistics operation.

See GetStatistics for more information about the operation and its arguments.

pub fn get_topic_rule(&self) -> GetTopicRule<C, M, R>

Constructs a fluent builder for the GetTopicRule operation.

See GetTopicRule for more information about the operation and its arguments.

pub fn get_topic_rule_destination(&self) -> GetTopicRuleDestination<C, M, R>

Constructs a fluent builder for the GetTopicRuleDestination operation.

See GetTopicRuleDestination for more information about the operation and its arguments.

pub fn get_v2_logging_options(&self) -> GetV2LoggingOptions<C, M, R>

Constructs a fluent builder for the GetV2LoggingOptions operation.

See GetV2LoggingOptions for more information about the operation and its arguments.

pub fn list_active_violations(&self) -> ListActiveViolations<C, M, R>

Constructs a fluent builder for the ListActiveViolations operation.

See ListActiveViolations for more information about the operation and its arguments.

pub fn list_attached_policies(&self) -> ListAttachedPolicies<C, M, R>

Constructs a fluent builder for the ListAttachedPolicies operation.

See ListAttachedPolicies for more information about the operation and its arguments.

pub fn list_audit_findings(&self) -> ListAuditFindings<C, M, R>

Constructs a fluent builder for the ListAuditFindings operation.

See ListAuditFindings for more information about the operation and its arguments.

pub fn list_audit_mitigation_actions_executions(
&self
) -> ListAuditMitigationActionsExecutions<C, M, R>

Constructs a fluent builder for the ListAuditMitigationActionsExecutions operation.

See ListAuditMitigationActionsExecutions for more information about the operation and its arguments.

pub fn list_audit_mitigation_actions_tasks(
&self
) -> ListAuditMitigationActionsTasks<C, M, R>

Constructs a fluent builder for the ListAuditMitigationActionsTasks operation.

See ListAuditMitigationActionsTasks for more information about the operation and its arguments.

pub fn list_audit_suppressions(&self) -> ListAuditSuppressions<C, M, R>

Constructs a fluent builder for the ListAuditSuppressions operation.

See ListAuditSuppressions for more information about the operation and its arguments.

pub fn list_audit_tasks(&self) -> ListAuditTasks<C, M, R>

Constructs a fluent builder for the ListAuditTasks operation.

See ListAuditTasks for more information about the operation and its arguments.

pub fn list_authorizers(&self) -> ListAuthorizers<C, M, R>

Constructs a fluent builder for the ListAuthorizers operation.

See ListAuthorizers for more information about the operation and its arguments.

pub fn list_billing_groups(&self) -> ListBillingGroups<C, M, R>

Constructs a fluent builder for the ListBillingGroups operation.

See ListBillingGroups for more information about the operation and its arguments.

pub fn list_ca_certificates(&self) -> ListCACertificates<C, M, R>

Constructs a fluent builder for the ListCACertificates operation.

See ListCACertificates for more information about the operation and its arguments.

pub fn list_certificates(&self) -> ListCertificates<C, M, R>

Constructs a fluent builder for the ListCertificates operation.

See ListCertificates for more information about the operation and its arguments.

pub fn list_certificates_by_ca(&self) -> ListCertificatesByCA<C, M, R>

Constructs a fluent builder for the ListCertificatesByCA operation.

See ListCertificatesByCA for more information about the operation and its arguments.

pub fn list_custom_metrics(&self) -> ListCustomMetrics<C, M, R>

Constructs a fluent builder for the ListCustomMetrics operation.

See ListCustomMetrics for more information about the operation and its arguments.

pub fn list_detect_mitigation_actions_executions(
&self
) -> ListDetectMitigationActionsExecutions<C, M, R>

Constructs a fluent builder for the ListDetectMitigationActionsExecutions operation.

See ListDetectMitigationActionsExecutions for more information about the operation and its arguments.

pub fn list_detect_mitigation_actions_tasks(
&self
) -> ListDetectMitigationActionsTasks<C, M, R>

Constructs a fluent builder for the ListDetectMitigationActionsTasks operation.

See ListDetectMitigationActionsTasks for more information about the operation and its arguments.

pub fn list_dimensions(&self) -> ListDimensions<C, M, R>

Constructs a fluent builder for the ListDimensions operation.

See ListDimensions for more information about the operation and its arguments.

pub fn list_domain_configurations(&self) -> ListDomainConfigurations<C, M, R>

Constructs a fluent builder for the ListDomainConfigurations operation.

See ListDomainConfigurations for more information about the operation and its arguments.

pub fn list_fleet_metrics(&self) -> ListFleetMetrics<C, M, R>

Constructs a fluent builder for the ListFleetMetrics operation.

See ListFleetMetrics for more information about the operation and its arguments.

pub fn list_indices(&self) -> ListIndices<C, M, R>

Constructs a fluent builder for the ListIndices operation.

See ListIndices for more information about the operation and its arguments.

pub fn list_job_executions_for_job(&self) -> ListJobExecutionsForJob<C, M, R>

Constructs a fluent builder for the ListJobExecutionsForJob operation.

See ListJobExecutionsForJob for more information about the operation and its arguments.

pub fn list_job_executions_for_thing(
&self
) -> ListJobExecutionsForThing<C, M, R>

Constructs a fluent builder for the ListJobExecutionsForThing operation.

See ListJobExecutionsForThing for more information about the operation and its arguments.

pub fn list_jobs(&self) -> ListJobs<C, M, R>

Constructs a fluent builder for the ListJobs operation.

See ListJobs for more information about the operation and its arguments.

pub fn list_job_templates(&self) -> ListJobTemplates<C, M, R>

Constructs a fluent builder for the ListJobTemplates operation.

See ListJobTemplates for more information about the operation and its arguments.

pub fn list_mitigation_actions(&self) -> ListMitigationActions<C, M, R>

Constructs a fluent builder for the ListMitigationActions operation.

See ListMitigationActions for more information about the operation and its arguments.

pub fn list_ota_updates(&self) -> ListOTAUpdates<C, M, R>

Constructs a fluent builder for the ListOTAUpdates operation.

See ListOTAUpdates for more information about the operation and its arguments.

pub fn list_outgoing_certificates(&self) -> ListOutgoingCertificates<C, M, R>

Constructs a fluent builder for the ListOutgoingCertificates operation.

See ListOutgoingCertificates for more information about the operation and its arguments.

pub fn list_policies(&self) -> ListPolicies<C, M, R>

Constructs a fluent builder for the ListPolicies operation.

See ListPolicies for more information about the operation and its arguments.

pub fn list_policy_principals(&self) -> ListPolicyPrincipals<C, M, R>

Constructs a fluent builder for the ListPolicyPrincipals operation.

See ListPolicyPrincipals for more information about the operation and its arguments.

pub fn list_policy_versions(&self) -> ListPolicyVersions<C, M, R>

Constructs a fluent builder for the ListPolicyVersions operation.

See ListPolicyVersions for more information about the operation and its arguments.

pub fn list_principal_policies(&self) -> ListPrincipalPolicies<C, M, R>

Constructs a fluent builder for the ListPrincipalPolicies operation.

See ListPrincipalPolicies for more information about the operation and its arguments.

pub fn list_principal_things(&self) -> ListPrincipalThings<C, M, R>

Constructs a fluent builder for the ListPrincipalThings operation.

See ListPrincipalThings for more information about the operation and its arguments.

pub fn list_provisioning_templates(&self) -> ListProvisioningTemplates<C, M, R>

Constructs a fluent builder for the ListProvisioningTemplates operation.

See ListProvisioningTemplates for more information about the operation and its arguments.

pub fn list_provisioning_template_versions(
&self
) -> ListProvisioningTemplateVersions<C, M, R>

Constructs a fluent builder for the ListProvisioningTemplateVersions operation.

See ListProvisioningTemplateVersions for more information about the operation and its arguments.

pub fn list_role_aliases(&self) -> ListRoleAliases<C, M, R>

Constructs a fluent builder for the ListRoleAliases operation.

See ListRoleAliases for more information about the operation and its arguments.

pub fn list_scheduled_audits(&self) -> ListScheduledAudits<C, M, R>

Constructs a fluent builder for the ListScheduledAudits operation.

See ListScheduledAudits for more information about the operation and its arguments.

pub fn list_security_profiles(&self) -> ListSecurityProfiles<C, M, R>

Constructs a fluent builder for the ListSecurityProfiles operation.

See ListSecurityProfiles for more information about the operation and its arguments.

pub fn list_security_profiles_for_target(
&self
) -> ListSecurityProfilesForTarget<C, M, R>

Constructs a fluent builder for the ListSecurityProfilesForTarget operation.

See ListSecurityProfilesForTarget for more information about the operation and its arguments.

pub fn list_streams(&self) -> ListStreams<C, M, R>

Constructs a fluent builder for the ListStreams operation.

See ListStreams for more information about the operation and its arguments.

pub fn list_tags_for_resource(&self) -> ListTagsForResource<C, M, R>

Constructs a fluent builder for the ListTagsForResource operation.

See ListTagsForResource for more information about the operation and its arguments.

pub fn list_targets_for_policy(&self) -> ListTargetsForPolicy<C, M, R>

Constructs a fluent builder for the ListTargetsForPolicy operation.

See ListTargetsForPolicy for more information about the operation and its arguments.

pub fn list_targets_for_security_profile(
&self
) -> ListTargetsForSecurityProfile<C, M, R>

Constructs a fluent builder for the ListTargetsForSecurityProfile operation.

See ListTargetsForSecurityProfile for more information about the operation and its arguments.

pub fn list_thing_groups(&self) -> ListThingGroups<C, M, R>

Constructs a fluent builder for the ListThingGroups operation.

See ListThingGroups for more information about the operation and its arguments.

pub fn list_thing_groups_for_thing(&self) -> ListThingGroupsForThing<C, M, R>

Constructs a fluent builder for the ListThingGroupsForThing operation.

See ListThingGroupsForThing for more information about the operation and its arguments.

pub fn list_thing_principals(&self) -> ListThingPrincipals<C, M, R>

Constructs a fluent builder for the ListThingPrincipals operation.

See ListThingPrincipals for more information about the operation and its arguments.

pub fn list_thing_registration_task_reports(
&self
) -> ListThingRegistrationTaskReports<C, M, R>

Constructs a fluent builder for the ListThingRegistrationTaskReports operation.

See ListThingRegistrationTaskReports for more information about the operation and its arguments.

pub fn list_thing_registration_tasks(
&self
) -> ListThingRegistrationTasks<C, M, R>

Constructs a fluent builder for the ListThingRegistrationTasks operation.

See ListThingRegistrationTasks for more information about the operation and its arguments.

pub fn list_things(&self) -> ListThings<C, M, R>

Constructs a fluent builder for the ListThings operation.

See ListThings for more information about the operation and its arguments.

pub fn list_things_in_billing_group(&self) -> ListThingsInBillingGroup<C, M, R>

Constructs a fluent builder for the ListThingsInBillingGroup operation.

See ListThingsInBillingGroup for more information about the operation and its arguments.

pub fn list_things_in_thing_group(&self) -> ListThingsInThingGroup<C, M, R>

Constructs a fluent builder for the ListThingsInThingGroup operation.

See ListThingsInThingGroup for more information about the operation and its arguments.

pub fn list_thing_types(&self) -> ListThingTypes<C, M, R>

Constructs a fluent builder for the ListThingTypes operation.

See ListThingTypes for more information about the operation and its arguments.

pub fn list_topic_rule_destinations(&self) -> ListTopicRuleDestinations<C, M, R>

Constructs a fluent builder for the ListTopicRuleDestinations operation.

See ListTopicRuleDestinations for more information about the operation and its arguments.

pub fn list_topic_rules(&self) -> ListTopicRules<C, M, R>

Constructs a fluent builder for the ListTopicRules operation.

See ListTopicRules for more information about the operation and its arguments.

pub fn list_v2_logging_levels(&self) -> ListV2LoggingLevels<C, M, R>

Constructs a fluent builder for the ListV2LoggingLevels operation.

See ListV2LoggingLevels for more information about the operation and its arguments.

pub fn list_violation_events(&self) -> ListViolationEvents<C, M, R>

Constructs a fluent builder for the ListViolationEvents operation.

See ListViolationEvents for more information about the operation and its arguments.

pub fn put_verification_state_on_violation(
&self
) -> PutVerificationStateOnViolation<C, M, R>

Constructs a fluent builder for the PutVerificationStateOnViolation operation.

See PutVerificationStateOnViolation for more information about the operation and its arguments.

pub fn register_ca_certificate(&self) -> RegisterCACertificate<C, M, R>

Constructs a fluent builder for the RegisterCACertificate operation.

See RegisterCACertificate for more information about the operation and its arguments.

pub fn register_certificate(&self) -> RegisterCertificate<C, M, R>

Constructs a fluent builder for the RegisterCertificate operation.

See RegisterCertificate for more information about the operation and its arguments.

pub fn register_certificate_without_ca(
&self
) -> RegisterCertificateWithoutCA<C, M, R>

Constructs a fluent builder for the RegisterCertificateWithoutCA operation.

See RegisterCertificateWithoutCA for more information about the operation and its arguments.

pub fn register_thing(&self) -> RegisterThing<C, M, R>

Constructs a fluent builder for the RegisterThing operation.

See RegisterThing for more information about the operation and its arguments.

pub fn reject_certificate_transfer(&self) -> RejectCertificateTransfer<C, M, R>

Constructs a fluent builder for the RejectCertificateTransfer operation.

See RejectCertificateTransfer for more information about the operation and its arguments.

pub fn remove_thing_from_billing_group(
&self
) -> RemoveThingFromBillingGroup<C, M, R>

Constructs a fluent builder for the RemoveThingFromBillingGroup operation.

See RemoveThingFromBillingGroup for more information about the operation and its arguments.

pub fn remove_thing_from_thing_group(
&self
) -> RemoveThingFromThingGroup<C, M, R>

Constructs a fluent builder for the RemoveThingFromThingGroup operation.

See RemoveThingFromThingGroup for more information about the operation and its arguments.

pub fn replace_topic_rule(&self) -> ReplaceTopicRule<C, M, R>

Constructs a fluent builder for the ReplaceTopicRule operation.

See ReplaceTopicRule for more information about the operation and its arguments.

pub fn search_index(&self) -> SearchIndex<C, M, R>

Constructs a fluent builder for the SearchIndex operation.

See SearchIndex for more information about the operation and its arguments.

pub fn set_default_authorizer(&self) -> SetDefaultAuthorizer<C, M, R>

Constructs a fluent builder for the SetDefaultAuthorizer operation.

See SetDefaultAuthorizer for more information about the operation and its arguments.

pub fn set_default_policy_version(&self) -> SetDefaultPolicyVersion<C, M, R>

Constructs a fluent builder for the SetDefaultPolicyVersion operation.

See SetDefaultPolicyVersion for more information about the operation and its arguments.

pub fn set_logging_options(&self) -> SetLoggingOptions<C, M, R>

Constructs a fluent builder for the SetLoggingOptions operation.

See SetLoggingOptions for more information about the operation and its arguments.

pub fn set_v2_logging_level(&self) -> SetV2LoggingLevel<C, M, R>

Constructs a fluent builder for the SetV2LoggingLevel operation.

See SetV2LoggingLevel for more information about the operation and its arguments.

pub fn set_v2_logging_options(&self) -> SetV2LoggingOptions<C, M, R>

Constructs a fluent builder for the SetV2LoggingOptions operation.

See SetV2LoggingOptions for more information about the operation and its arguments.

pub fn start_audit_mitigation_actions_task(
&self
) -> StartAuditMitigationActionsTask<C, M, R>

Constructs a fluent builder for the StartAuditMitigationActionsTask operation.

See StartAuditMitigationActionsTask for more information about the operation and its arguments.

pub fn start_detect_mitigation_actions_task(
&self
) -> StartDetectMitigationActionsTask<C, M, R>

Constructs a fluent builder for the StartDetectMitigationActionsTask operation.

See StartDetectMitigationActionsTask for more information about the operation and its arguments.

pub fn start_on_demand_audit_task(&self) -> StartOnDemandAuditTask<C, M, R>

Constructs a fluent builder for the StartOnDemandAuditTask operation.

See StartOnDemandAuditTask for more information about the operation and its arguments.

pub fn start_thing_registration_task(
&self
) -> StartThingRegistrationTask<C, M, R>

Constructs a fluent builder for the StartThingRegistrationTask operation.

See StartThingRegistrationTask for more information about the operation and its arguments.

pub fn stop_thing_registration_task(&self) -> StopThingRegistrationTask<C, M, R>

Constructs a fluent builder for the StopThingRegistrationTask operation.

See StopThingRegistrationTask for more information about the operation and its arguments.

pub fn tag_resource(&self) -> TagResource<C, M, R>

Constructs a fluent builder for the TagResource operation.

See TagResource for more information about the operation and its arguments.

pub fn test_authorization(&self) -> TestAuthorization<C, M, R>

Constructs a fluent builder for the TestAuthorization operation.

See TestAuthorization for more information about the operation and its arguments.

pub fn test_invoke_authorizer(&self) -> TestInvokeAuthorizer<C, M, R>

Constructs a fluent builder for the TestInvokeAuthorizer operation.

See TestInvokeAuthorizer for more information about the operation and its arguments.

pub fn transfer_certificate(&self) -> TransferCertificate<C, M, R>

Constructs a fluent builder for the TransferCertificate operation.

See TransferCertificate for more information about the operation and its arguments.

pub fn untag_resource(&self) -> UntagResource<C, M, R>

Constructs a fluent builder for the UntagResource operation.

See UntagResource for more information about the operation and its arguments.

pub fn update_account_audit_configuration(
&self
) -> UpdateAccountAuditConfiguration<C, M, R>

Constructs a fluent builder for the UpdateAccountAuditConfiguration operation.

See UpdateAccountAuditConfiguration for more information about the operation and its arguments.

pub fn update_audit_suppression(&self) -> UpdateAuditSuppression<C, M, R>

Constructs a fluent builder for the UpdateAuditSuppression operation.

See UpdateAuditSuppression for more information about the operation and its arguments.

pub fn update_authorizer(&self) -> UpdateAuthorizer<C, M, R>

Constructs a fluent builder for the UpdateAuthorizer operation.

See UpdateAuthorizer for more information about the operation and its arguments.

pub fn update_billing_group(&self) -> UpdateBillingGroup<C, M, R>

Constructs a fluent builder for the UpdateBillingGroup operation.

See UpdateBillingGroup for more information about the operation and its arguments.

pub fn update_ca_certificate(&self) -> UpdateCACertificate<C, M, R>

Constructs a fluent builder for the UpdateCACertificate operation.

See UpdateCACertificate for more information about the operation and its arguments.

pub fn update_certificate(&self) -> UpdateCertificate<C, M, R>

Constructs a fluent builder for the UpdateCertificate operation.

See UpdateCertificate for more information about the operation and its arguments.

pub fn update_custom_metric(&self) -> UpdateCustomMetric<C, M, R>

Constructs a fluent builder for the UpdateCustomMetric operation.

See UpdateCustomMetric for more information about the operation and its arguments.

pub fn update_dimension(&self) -> UpdateDimension<C, M, R>

Constructs a fluent builder for the UpdateDimension operation.

See UpdateDimension for more information about the operation and its arguments.

pub fn update_domain_configuration(&self) -> UpdateDomainConfiguration<C, M, R>

Constructs a fluent builder for the UpdateDomainConfiguration operation.

See UpdateDomainConfiguration for more information about the operation and its arguments.

pub fn update_dynamic_thing_group(&self) -> UpdateDynamicThingGroup<C, M, R>

Constructs a fluent builder for the UpdateDynamicThingGroup operation.

See UpdateDynamicThingGroup for more information about the operation and its arguments.

pub fn update_event_configurations(&self) -> UpdateEventConfigurations<C, M, R>

Constructs a fluent builder for the UpdateEventConfigurations operation.

See UpdateEventConfigurations for more information about the operation and its arguments.

pub fn update_fleet_metric(&self) -> UpdateFleetMetric<C, M, R>

Constructs a fluent builder for the UpdateFleetMetric operation.

See UpdateFleetMetric for more information about the operation and its arguments.

pub fn update_indexing_configuration(
&self
) -> UpdateIndexingConfiguration<C, M, R>

Constructs a fluent builder for the UpdateIndexingConfiguration operation.

See UpdateIndexingConfiguration for more information about the operation and its arguments.

pub fn update_job(&self) -> UpdateJob<C, M, R>

Constructs a fluent builder for the UpdateJob operation.

See UpdateJob for more information about the operation and its arguments.

pub fn update_mitigation_action(&self) -> UpdateMitigationAction<C, M, R>

Constructs a fluent builder for the UpdateMitigationAction operation.

See UpdateMitigationAction for more information about the operation and its arguments.

pub fn update_provisioning_template(
&self
) -> UpdateProvisioningTemplate<C, M, R>

Constructs a fluent builder for the UpdateProvisioningTemplate operation.

See UpdateProvisioningTemplate for more information about the operation and its arguments.

pub fn update_role_alias(&self) -> UpdateRoleAlias<C, M, R>

Constructs a fluent builder for the UpdateRoleAlias operation.

See UpdateRoleAlias for more information about the operation and its arguments.

pub fn update_scheduled_audit(&self) -> UpdateScheduledAudit<C, M, R>

Constructs a fluent builder for the UpdateScheduledAudit operation.

See UpdateScheduledAudit for more information about the operation and its arguments.

pub fn update_security_profile(&self) -> UpdateSecurityProfile<C, M, R>

Constructs a fluent builder for the UpdateSecurityProfile operation.

See UpdateSecurityProfile for more information about the operation and its arguments.

pub fn update_stream(&self) -> UpdateStream<C, M, R>

Constructs a fluent builder for the UpdateStream operation.

See UpdateStream for more information about the operation and its arguments.

pub fn update_thing(&self) -> UpdateThing<C, M, R>

Constructs a fluent builder for the UpdateThing operation.

See UpdateThing for more information about the operation and its arguments.

pub fn update_thing_group(&self) -> UpdateThingGroup<C, M, R>

Constructs a fluent builder for the UpdateThingGroup operation.

See UpdateThingGroup for more information about the operation and its arguments.

pub fn update_thing_groups_for_thing(
&self
) -> UpdateThingGroupsForThing<C, M, R>

Constructs a fluent builder for the UpdateThingGroupsForThing operation.

See UpdateThingGroupsForThing for more information about the operation and its arguments.

pub fn update_topic_rule_destination(
&self
) -> UpdateTopicRuleDestination<C, M, R>

Constructs a fluent builder for the UpdateTopicRuleDestination operation.

See UpdateTopicRuleDestination for more information about the operation and its arguments.

pub fn validate_security_profile_behaviors(
&self
) -> ValidateSecurityProfileBehaviors<C, M, R>

Constructs a fluent builder for the ValidateSecurityProfileBehaviors operation.

See ValidateSecurityProfileBehaviors for more information about the operation and its arguments.

impl<C> Client<C, AwsMiddleware, Standard>

pub fn from_conf_conn(conf: Config, conn: C) -> Self

Creates a client with the given service config and connector override.

impl Client<DynConnector, AwsMiddleware, Standard>

pub fn new(config: &Config) -> Self

Creates a new client from a shared config.

pub fn from_conf(conf: Config) -> Self

Creates a new client from the service Config.